### 2022-02-01 #4116 @whedon
[PRE REVIEW]: E2EDNA 2.0: Python Pipeline for Simulating DNA Aptamers with Ligands
**Submitting author:** @taoliu032 (<a href="http://orcid.org/0000-0002-1082-5570">Tao Liu</a>)
**Repository:** <a href="https://github.com/siminegroup/E2EDNA2" target ="_blank">https://github.com/siminegroup/E2EDNA2</a>
**Version:** v1.0
