### 2022-02-09 #4116 @csoneson
👋🏻 @taoliu032 - I had a quick look at the submission, and I was wondering if you could perhaps quickly clarify a couple of things before we move on:

* The relation to E2EDNA ([repo](https://github.com/InfluenceFunctional/E2EDNA), [paper](https://pubs.acs.org/doi/10.1021/acs.jcim.1c00696)) - admittedly this is not really my field of expertise, but it wasn't totally clear to me. The paper mentions that E2EDNA 2.0 is more user-friendly and easier to install, but it wasn't obvious if there are other significant additions/advantages/differences, and whether you would recommend E2EDNA users to switch to E2EDNA 2.0, or if they are to some extent complementary and you plan to maintain both going forward (the paper denotes them as predecessor/successor, but I couldn't see e.g. a note in the E2EDNA repo pointing people to E2EDNA 2.0).
* I see only instructions for installation on macOS, and a file `macos_installation.sh` in the repo - do you expect it to work also for other operating systems, or is there an inherent restriction to macOS? I suspect the provided environment yml file will not be portable.

Thanks!
