### 2022-02-11 b2cccbe
Update EXAMPLE_SUB_cluster.sh
