+++ EXAMPLE_SUB_cluster.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #SBATCH --account=ACCOUNT
-#SBATCH --gres=gpu:v100:1
+#SBATCH --gres=gpu:a100:1
 #SBATCH --cpus-per-task=1
 #SBATCH --mem=4G
 #SBATCH --time=0-4:00
