### 2022-02-13 #4116 @csoneson
👋🏻 @ayush9pandey, @ayaanhossain - would you be interested in reviewing this submission for JOSS?

> E2EDNA 2.0: Python Pipeline for Simulating DNA Aptamers with Ligands
