### 2022-02-15 #4116 @csoneson
👋🏻 @JoaoRodrigues, @rvhonorato - would you be interested in reviewing this submission for JOSS?

> E2EDNA 2.0: Python Pipeline for Simulating DNA Aptamers with Ligands
