### 2022-02-16 #4116 @taoliu032
@csoneson May I ask to slightly modify the code? One of the toggles was not perfectly set up but it does not affect the backbone or main features of the code at all. I literally just need to edit one line and delete a few.

This is my first ever submission to JOSS (very exciting experience!), so I am not sure if I could make changes to the code at the moment or if I should wait till the official review starts to bring up this request. I wonder if you could point out what action would be deemed to be most appropriate?

My apologies for any inconvenience! Thank you very much!
