### 2022-02-16 8b475ac
Fix the toggle in main.py to pick up from chk file
