### 2022-02-18 #4116 @csoneson
Pinging @JoaoRodrigues and @ayaanhossain - would one of you be interested in reviewing this submission for JOSS?

> E2EDNA 2.0: Python Pipeline for Simulating DNA Aptamers with Ligands
