### 2022-02-18 #4182 @whedon
[REVIEW]: E2EDNA 2.0: Python Pipeline for Simulating DNA Aptamers with Ligands
**Submitting author:** @taoliu032 (<a href="http://orcid.org/0000-0002-1082-5570">Tao Liu</a>)
**Repository:** https://github.com/siminegroup/E2EDNA2
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v2.0.0
