### 2022-02-21 #4182 @rvhonorato
This code might have a licensing issue, but since **i am not a licensing expert**, I will raise the point here for discussion:

I have noticed in your installation instructions the following:

> params['ld setup path'], params['ld run path'], etc: paths to various [LightDock](https://lightdock.org/) Python scripts, providing they are in a working directory. No need to change these paths. The paths are used to automatically copy the scripts to ld_scripts/ within a working directory from [lib/lightdock/](https://github.com/InfluenceFunctional/OpenDNA/tree/main/lib/lightdock).

Which then, as expected results in the lightdock scripts being **copied** from the lightdock package and **in** your execution.

![Screenshot 2022-02-21 at 13 37 19](https://user-images.githubusercontent.com/9445814/154956607-fda88b44-2794-4c58-8530-33999f4eb88c.png)

To my understanding the fact the the script are being **copied** rather than **imported** the lightdock from the library makes necessary for **your** code to also be GPLv3, whereas it is now Apache-2.

Since you are just using the scripts and not modifying them (as far as I know) it is better to import the scripts and use then from within your code to avoid any issues.

Ping brianjimenez (lightdock creator) and also @JoaoRodrigues (the other reviewer) since they might also have some insights into this.
