### 2022-02-21 #4182 @brianjimenez
Thanks for the ping @rvhonorato
Indeed, it could be a license conflict here since redistributing code with a different license than LightDock (GPLv3) might not be allowed (but needs a double check for Apache-2 + GPLv3 compatibility). If authors are not modifying the code, simpler could be just to add LightDock as a pip dependency and those scripts would be available in the path.
