### 2022-02-21 #4182 @taoliu032
@rvhonorato Thank you so much for raising the concern! We are not very experienced in choosing a license for open source software -- I was in fact planning to consult JOSS reviewers on the choice of license :) But certainly we do not want to have any licensing conflict.

@brianjimenez Thank you for helping out! We got the scripts from https://github.com/lightdock/lightdock/tree/0.9.2/bin and we did include LightDock package using pip when creating a conda virtual environment; Otherwise, those scripts would not run either.
I will try the direct import for the methods in those scripts. One (probably unnecessary) concern is that: when I looked at the "lightdock" folder under "site-packages" in the conda environment, I did not see those scripts. I will check the LightDock documentation to figure out how to do the import properly. If you can point out any direction or document, it would be super helpful too. Thank you again!
