### 2022-02-21 #4182 @brianjimenez
Hi @taoliu032 , glad to help! To find LightDock scripts, I assume you use `virtualenv` command (conda should behave in a similar way):

```bash
virtualenv venv
source venv/bin/activate
pip install lightdock
```

After installation ends, you should find LightDock scripts on `venv/bin` folder. Try also `which ant_thony.py` to locate scripts. When the virtual environment (or conda) is activated, all LightDock scripts are available at any path.

Hope it helps!
