### 2022-02-22 #4182 @taoliu032
@brianjimenez - I removed the commands of copying individual lightdock scripts and our code can still run successfully. Thank you for the suggestion!

@rvhonorato - After removing the script-copying command, I have tested the code both on my local macOS machine and our Linux-based computer cluster (the job waited in the queue for a while). Both tests ran successfully and I have also pushed the updates to the JOSS branch :-) Please feel free to download the new code - sorry for causing any inconvenience! Let me know if there is any other concern or questions, thank you!
