### 2022-02-22 7ec6f12
Delete lib/lightdock directory
Once the virtual environment (with LightDock package installed) is active, these scripts are already available at any path.
