### 2022-02-22 8ed7a60
Update the usage of lightdock Python scripts
No need to explicitly copy the lightdock Python scripts to the working directory. Once the virtual environment (with LightDock package installed) is active, those scripts are already available at any path.
