### 2022-03-08 70e956a
Update variable names for ld script to avoid conflict
In interfaces.py: previously, variable names like "self.run" can conflict with the method "run" within the ld class. So change "self.run" to "self.runScript" to distinguish them from the methods.
