+++ main.py
@@ -21,32 +21,131 @@ Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
    See the License for the specific language governing permissions and
    limitations under the License.
 '''
+import argparse
 from opendna import *
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
+
+def get_args():
+    ''' Parse commandline arguments '''
+
+    parser = argparse.ArgumentParser(
+        description='E2EDNA: Simulate DNA aptamers complexed with target ligands',
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+
+    run_info = parser.add_argument_group('Run Parameters')
+    system_info = parser.add_argument_group('System Configuration')
+    paths = parser.add_argument_group('Directory Settings')
+
+    run_info.add_argument('-o',
+                             '--outdir',
+                             metavar='DIR',
+                             type=str,
+                             required=True,
+                             help='Output directory for run created in --workdir')
+    run_info.add_argument('-m',
+                             '--mode',
+                             metavar='MODE',
+                             type=str,
+                             required=True,
+                             help='Run mode')
+    run_info.add_argument('-a',
+                             '--aptamer',
+                             metavar='SEQ',
+                             type=str,
+                             required=True,
+                             help="DNA Aptamer sequence (5'-3')")
+    run_info.add_argument('-l',
+                             '--ligand',
+                             metavar='PDB',
+                             type=argparse.FileType('rt'),
+                             required=True,
+                             help='Filename of target ligand')
+    run_info.add_argument('-t',
+                             '--ligand_type',
+                             metavar='TYPE',
+                             type=str,
+                             default='',
+                             help='Target ligand molecule type',
+                             choices=['peptide', 'DNA', 'RNA', 'other'])
+    run_info.add_argument('-s',
+                             '--ligand_sequence',
+                             metavar='SEQ',
+                             type=str,
+                             default='',
+                             help='Target ligand sequence if peptide, DNA, or RNA',
+                             choices=['peptide', 'DNA', 'RNA', 'other'])
+    
+
+    system_info.add_argument('-d',
+                             '--device',
+                             metavar='RUN',
+                             type=str,
+                             default='local',
+                             help='Device configuration',
+                             choices=['local', 'cluster'])
+    system_info.add_argument('-os',
+                             '--system',
+                             metavar='OS',
+                             type=str,
+                             default='macos',
+                             help='Operating system',
+                             choices=['macos', 'linux', 'WSL'])
+    system_info.add_argument('-p',
+                             '--platform',
+                             metavar='DEV',
+                             type=str,
+                             default='CPU',
+                             help='Processing platform',
+                             choices=['CPU', 'CUDA'])
+
+    paths.add_argument('-d',
+                          '--workdir',
+                          metavar='DIR',
+                          type=str,
+                          default='./localruns',
+                          help='Output directory to store runs')
+    paths.add_argument('-md',
+                          '--mmb_dir',
+                          metavar='DIR',
+                          type=str,
+                          default='Installer.*/lib',
+                          help='MMB library directory')
+    paths.add_argument('-mb',
+                          '--mmb',
[…]
