### 2022-03-08 044d903
begin validating arguments
