+++ main.py
@@ -22,7 +22,9 @@ Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
    limitations under the License.
 '''
 import argparse
+import glob
 from opendna import *
+import os
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
@@ -59,7 +61,7 @@ def get_args():
     run_info.add_argument('-l',
                              '--ligand',
                              metavar='PDB',
-                             type=argparse.FileType('rt'),
+                             type=str,
                              required=True,
                              help='Filename of target ligand')
     run_info.add_argument('-t',
@@ -68,14 +70,17 @@ def get_args():
                              type=str,
                              default='',
                              help='Target ligand molecule type',
-                             choices=['peptide', 'DNA', 'RNA', 'other'])
+                             choices=['peptide', 'DNA', 'RNA', 'other', 'False'])
     run_info.add_argument('-s',
-                             '--ligand_sequence',
+                             '--ligand_seq',
                              metavar='SEQ',
                              type=str,
                              default='',
-                             help='Target ligand sequence if peptide, DNA, or RNA',
-                             choices=['peptide', 'DNA', 'RNA', 'other'])
+                             help='Target ligand sequence if peptide, DNA, or RNA')
+    run_info.add_argument('-f',
+                             '--force',
+                             action='store_true',
+                             help='Overwrite existing --outdir')
     
 
     system_info.add_argument('-d',
@@ -100,7 +105,7 @@ def get_args():
                              help='Processing platform',
                              choices=['CPU', 'CUDA'])
 
-    paths.add_argument('-d',
+    paths.add_argument('-w',
                           '--workdir',
                           metavar='DIR',
                           type=str,
@@ -110,16 +115,74 @@ def get_args():
                           '--mmb_dir',
                           metavar='DIR',
                           type=str,
-                          default='Installer.*/lib',
+                          default='Installer-*/lib',
                           help='MMB library directory')
     paths.add_argument('-mb',
                           '--mmb',
                           metavar='MMB',
                           type=str,
-                          default='Installer.*/bin/MMB',
+                          default='Installer-*/bin/MMB',
                           help='Path to MMB executable')
 
     args = parser.parse_args()
+
+    # Check if workdir exists, and create if not
+    if not os.path.isdir(args.workdir):
+        os.mkdir(args.workdir)
+
+    # Check if output directory for run exists;
+    # Either fail or force overwrite if so
+    out_dir = os.path.join(args.workdir, args.outdir)
+    if os.path.isdir(out_dir):
+        if args.force:
+            os.removedirs()
+        else:
+            parser.error(f'--outdir already exists at {out_dir}\n'
+                        '\t use -f/--force to overwrite')
+
+    # Validate run mode
+    valid_modes = ['2d structure', '3d coarse', '3d smooth',
+                   'coarse dock', 'smooth dock', 'free aptamer',
+                   'full dock', 'full binding']
+
+    if args.mode not in valid_modes:
+        parser.error(f'Invalid --mode "{args.mode}". Must be one of: ' +
+                     ', '.join(valid_modes) + '.')
+
+    # Read in aptamer sequence if it is in a file
+    if os.path.isfile(args.aptamer):
+        args.aptamer = open(args.aptamer).read().rstrip()
+    
[…]
