### 2022-03-08 aaaed06
improve directory handling
