### 2022-03-08 4f0bc5c
update readme to reflect argparse implementation
