+++ README.MD
@@ -1,7 +1,7 @@
 # Documentation
 ## E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
 
-### An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
+### An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules)../m  
 
 **Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine**
 
@@ -18,36 +18,84 @@ This installation path has been tested on macOS and it relies on conda and pip p
    * To fix this issue, please, first remove the installed environment by ``conda remove --name e2edna --all`` in command line, then either replace the conda activation command in the script with `source activate /path-to-envs/e2edna`, e.g., `source activate ~/miniconda3/envs/e2edna` and run the script again
    * OR run the commands one by one interactively in your terminal. 
 5. Register and download MMB from https://simtk.org/projects/rnatoolbox. Place the "Installer.*version.OS*" folder, such as "Installer.3_0.OSX" into the E2EDNA2 directory. N.B.: Do not specify `DYLD_LIBRARY_PATH` for macOS against the recommendations of the MMB installation guide. This is to avoid interference with the OpenMM module.
-6. Update 3 paths in [main.py](https://github.com/siminegroup/E2EDNA2/blob/main/main.py):
+
+## 2. Usage
+
+The usage and help statements can be accessed with the `-h/--help` flags:
+
 ```
-params['workdir'] = '/path-to-E2EDNA2-main/localruns'                                # directory manually created to store all future jobs
-params['mmb dir'] = '/path-to-E2EDNA2-main/Installer.version.OS/lib'                 # path to MMB dylib files
-params['mmb']     = '/path-to-E2EDNA2-main/Installer.version.OS/bin/MMB-executable'  # path to MMB executable
+$ ./main.py --help
+usage: main.py [-h] [-f] -r INT -m MODE -a SEQ -l PDB [-t TYPE] [-s SEQ]
+               [-d RUN] [-p DEV] [-w DIR] [-md DIR] [-mb MMB]
+
+E2EDNA: Simulate DNA aptamers complexed with target ligands
+
+optional arguments:
+  -h, --help            show this help message and exit
+  -f, --force           Overwrite existing --run_num (default: False)
+
+Run Parameters:
+  -r INT, --run_num INT
+                        Run number. Output will be written to
+                        {--workdir}/run{--run_num}/ (default: None)
+  -m MODE, --mode MODE  Run mode (default: None)
+  -a SEQ, --aptamer SEQ
+                        DNA Aptamer sequence (5'-3') (default: None)
+  -l PDB, --ligand PDB  Filename of target ligand (default: None)
+  -t TYPE, --ligand_type TYPE
+                        Target ligand molecule type (default: )
+  -s SEQ, --ligand_seq SEQ
+                        Target ligand sequence if peptide, DNA, or RNA
+                        (default: )
+
+System Configuration:
+  -d RUN, --device RUN  Device configuration (default: local)
+  -p DEV, --platform DEV
+                        Processing platform (default: CPU)
+
+Directory Settings:
+  -w DIR, --workdir DIR
+                        Output directory to store runs (default: ./localruns)
+  -md DIR, --mmb_dir DIR
+                        MMB library directory (default: Installer-*/lib)
+  -mb MMB, --mmb MMB    Path to MMB executable (default: Installer-*/bin/MMB)
 ```
-## 2. Running a job
-### Quickstart
-* Set 3 paths in `params`, as indicated in step6 of [Installation](https://github.com/siminegroup/E2EDNA2#1-installation) section. Leave the other keys of `params` unchanged.
[…]
