+++ automate_tests.sh
@@ -7,20 +7,20 @@ printf "Start automating tests one by one...\n"
 printf "===================================="
 
 printf "\nTESTING MODE #1: '2d structure'\n"
-python main.py --outdir 1 --mode '2d structure' --aptamer 'TAATGTTAATTG' --ligand 'False' --ligand_type '' --ligand_seq '' #> run1.out
+python main.py  --run_num 1 --mode '2d structure' --aptamer 'TAATGTTAATTG' --ligand 'False' --ligand_type '' --ligand_seq '' #> run1.out
 printf "\nEND OF TEST #1. Results are saved to folder \"run1\", where:\n"
 printf "\t2d structure: in record.txt\n"
 printf "===================================="
 
 printf "\nTESTING MODE #2: '3d coarse'\n"
-python main.py --outdir 2 --mode '3d coarse' --aptamer 'TAATGTTAATTG' --ligand 'False' --ligand_type '' --ligand_seq '' #> run2.out
+python main.py  --run_num 2 --mode '3d coarse' --aptamer 'TAATGTTAATTG' --ligand 'False' --ligand_type '' --ligand_seq '' #> run2.out
 printf "\nEND OF TEST #2. Results are saved to folder \"run2\", where:\n"
 printf "\t2d structure: in record.txt\n"
 printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
 printf "===================================="
 
 printf "\nTESTING MODE #3: '3d smooth'\n"
-python main.py --outdir 3 --mode '3d smooth' --aptamer 'TAATGTTAATTG' --ligand 'False' --ligand_type '' --ligand_seq '' #> run3.out
+python main.py  --run_num 3 --mode '3d smooth' --aptamer 'TAATGTTAATTG' --ligand 'False' --ligand_type '' --ligand_seq '' #> run3.out
 printf "\nEND OF TEST #3. Results are saved to folder \"run3\", where:\n"
 printf "\t2d structure: in record.txt\n"
 printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
@@ -29,7 +29,7 @@ printf "\tRelaxed aptamer structure: relaxedAptamer_0.pdb\n"
 printf "===================================="
 
 printf "\nTESTING MODE #4: 'coarse dock'\n"
-python main.py --outdir 4 --mode 'coarse dock' --aptamer 'TAATGTTAATTG' --ligand 'YQTQ.pdb' --ligand_type 'peptide' --ligand_seq 'YQTQTNSPRRAR' #> run4.out
+python main.py  --run_num 4 --mode 'coarse dock' --aptamer 'TAATGTTAATTG' --ligand 'YQTQ.pdb' --ligand_type 'peptide' --ligand_seq 'YQTQTNSPRRAR' #> run4.out
 printf "\nEND OF TEST #4. Results are saved to folder \"run4\", where:\n"
 printf "\t2d structure: in record.txt\n"
 printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
@@ -37,7 +37,7 @@ printf "\tMMB-folded aptamer docked by target ligand: complex_0_0.pdb (if dockin
 printf "===================================="
 
 printf "\nTESTING MODE #5: 'smooth dock'\n"
-python main.py --outdir 5 --mode 'smooth dock' --aptamer 'TAATGTTAATTG' --ligand 'YQTQ.pdb' --ligand_type 'peptide' --ligand_seq 'YQTQTNSPRRAR' #> run5.out
[…]
