### 2022-03-08 40ef553
broaden default mmb location flag wildcards
