### 2022-03-09 3d3fe1a
remove old command line parsing
