### 2022-03-09 aa75cbc
use argparse to validate --mode
