### 2022-03-09 siminegroup/E2EDNA2#14 @JoaoRodrigues
JOSS Review
Hi all,

Thanks for the invitation to review and congrats on the submission.

The general idea behind this submission is sound, and follows-up on a 2021 publication from the same authors on E2EDNA v1.0, published in JCIM. From my understanding, the code is essentially a re-write to use OpenMM instead of Tinker as the MD engine. While this is valuable - makes it simpler to install/run - the authors do not realize, in my opinion, this change to its fullest potential. The authors repository is not so much a "package" in the traditional sense, but more of a collection of scripts that automate a certain rigid protocol. I would rather see for instance, NUPACK being an optional dependency - as  a user, I could simply provide my own DNA molecules instead of being forced to use NUPACK. In this sense, I think this repository could use more work to stand out on its own compared to last year's publication.

In addition to this comments, I have a general comment on the repository itself. The authors should take some time to clean up files that are no longer useful for the protocol or that are simply part of the development workflow. Folders named `old`, or IDE config folders (`.idea`) should not be part of a published version of the repository, specially when they are even marked to be ignored in the `.gitignore` file. Same with the existence of both a `requirements.txt` file and an `environment.yml` file, whereas only the latter is used. As such, I believe that the authors should spend some time cleaning up the repository and setting up a more "traditional" structure to help potential users navigate through their code base more easily.

Further, I have a few starter questions about the manuscript, code and, licenses that I think should be clarified. Hopefully these will help the authors improve their work and repository/code.

### Licenses
- You're licensing the tool under the Apache license but you are including data (parameter sets) that falls under a difference license. In particular, I see the parameter files for the Amoeba forcefield taken from Tinker/OpenMM almost verbatim. Did you check with the appropriate developers if this sharing of the forcefield parameter files is allowed under their license, without any attribution?

### Installation
- The installation process is quite complex. As a user, I'd have to register and download NUPACK and MMB, as well as edit a series of files in order to get a functional installation. This is simply a suggestion for the developers to keep in mind.

- Related to the point above, have the authors considered using conda directly to install their software, instead of a custom shell script? `pdbfixer` is available as a conda package, and you could specify pip packages there too, e.g. `lightdock`. The installation could be reduced to a simple: 1) install nupack 2) install mmb 3) run `conda env create -f e2edna-env.yml`.

- On this last point, the authors should strip the granular version of the env `yaml` file otherwise conda will struggle with versions on anything but the authors' hardware.

- According to the README, the code is only tested on MacOS, although I'd imagine the most use would be on a compute cluster running Linux. Have the authors tried running their code on Linux?

### Misc
- In several sections of their documentation, the authors mention "OpenDNA". Was this the previous name of this package?
- It would be greatly beneficial for a user to have config files with installation paths, simulation settings etc, instead of having to edit source code. Would the authors be open to this change?

## Comments on the Manuscript
- In the "Statement of Need", the authors mention an "all-python" package several times. Being pedantic, this is not entirely true as their code relies on quite some compiled code in their dependencies (lightdock, openmm).
