### 2022-03-11 c1131b9
Clean env yaml and add lightdock to yaml
