### 2022-03-11 6a759e6
Update e2edna-env.yml: add pdbfixer package
Be aware of the nuisance between different versions of pdbfixer:
-- pdbfixer-1.7: "import simtk.opemm..."  => use this version on Computer Canada cluster Narval.
-- pdbfixer-1.8 and 1.8.1: "import openmm..." => use this version on any computer that installs openmm via conda (like in our e2edna-env.yml).
