+++ e2edna-env.yml
@@ -117,6 +117,7 @@ dependencies:
   - packmol=20.010
   - pandas=1.3.5
   - patsy=0.5.2
+  - pdbfixer=1.8.1
   - perl=5.32.1
   - pillow=8.4.0
   - pip=21.2.4
