### 2022-03-14 siminegroup/E2EDNA2#14 @taoliu032
Thanks very much for the comments.

(1) About comparison between E2EDNA (JCIM) and E2EDNA 2.0 (this manuscript):
> The general idea behind this submission is sound, and follows-up on a 2021 publication from the same authors on E2EDNA v1.0, published in JCIM. From my understanding, the code is essentially a re-write to use OpenMM instead of Tinker as the MD engine. While this is valuable - makes it simpler to install/run - the authors do not realize, in my opinion, this change to its fullest potential. The authors repository is not so much a "package" in the traditional sense, but more of a collection of scripts that automate a certain rigid protocol. I would rather see for instance, NUPACK being an optional dependency - as a user, I could simply provide my own DNA molecules instead of being forced to use NUPACK. In this sense, I think this repository could use more work to stand out on its own compared to last year's publication.
- Our response:
  - Compared to the original E2EDNA, we added more flexibility to E2EDNA 2.0 surrounding the 8 simulation modes. Specifically, we designed several "toggles"  on top of these modes to allow users to customize their simulation protocol depending on their demand. One of the toggles, `params['skip MMB']`, was designed exactly for the purpose the reviewer mentioned, which is to skip NUPACK and MMB if a user can provide a folded structure of DNA aptamer. Another example is `params['pick up from freeAptamerChk']` which allows a user to continue an unfinished MD sampling of free aptamer from a checkpoint file. This prevents a user from having to start the pipeline from scratch if a lengthy MD sampling is stopped for some reason.
  - We also would like to point out that we did not intend to provide every possible way of customizing the simulation pipeline, because there could be plenty of situations different users might encounter. We limited ourselves to developing a finite set of features for the common situations that we reasoned our potential users might run into and drew a line there for this submission. Meanwhile, we encourage anyone who is interested in making contributions to dig in our code, add features or improvement, as stated in the CONTRIBUTING.md ([link](https://github.com/siminegroup/E2EDNA2/blob/JOSS/CONTRIBUTING.md)). So far there has been some interest in doing so.

(2) About cleaning up the code base:
> In addition to this comments, I have a general comment on the repository itself. The authors should take some time to clean up files that are no longer useful for the protocol or that are simply part of the development workflow. Folders named `old`, or IDE config folders (`.idea`) should not be part of a published version of the repository, specially when they are even marked to be ignored in the `.gitignore` file. Same with the existence of both a `requirements.txt` file and an `environment.yml` file, whereas only the latter is used. As such, I believe that the authors should spend some time cleaning up the repository and setting up a more "traditional" structure to help potential users navigate through their code base more easily.
- Our response: Thanks for the suggestion to clean up the repository. We have done so in a branch called “JOSS” for this submission which contains the cleaned version of the code base and our manuscript. But unfortunately because of a misunderstanding the reviewer was directed to the wrong branch. Please take a look at the JOSS branch for a version that is up to the standards that the reviewer has requested.
- Changes to the code base: Unnecessary folders and files have been further deleted from the JOSS branch, which are `.idea`, `.gitignore` and `.gitattributes`.

(3) Licenses:
> * You're licensing the tool under the Apache license but you are including data (parameter sets) that falls under a difference license. In particular, I see the parameter files for the Amoeba forcefield taken from Tinker/OpenMM almost verbatim. Did you check with the appropriate developers if this sharing of the forcefield parameter files is allowed under their license, without any attribution?
- Our response:
  - The parameter sets mentioned by the reviewer are in the main branch, which is not meant to be used for this submission. In the JOSS branch (for this submission) which contains the cleaned version of the code base, all those parameter files were deleted therefore will not be shared if our submitted work gets published.
  - We chose Apache-2 license which does not include the condition of "same license" that GPLv3 does (shown in the [help website](https://choosealicense.com/licenses/) of choosing a license provided by GitHub). For this reason, I think that the Apache-2 license could give more flexibility to our software than GPLv3 as our E2EDNA (current and future versions) might be used by a larger software. But I am not a license expert and this is my own opinion. Meanwhile, we are in contact with the developer of LightDock (licensed under GPLv3) to work out if licensing E2EDNA 2.0 with Apache-2 would cause any conflict ([here](https://github.com/siminegroup/E2EDNA2/issues/15)). The other main external software packages used in the E2EDNA 2.0 (NUPACK, MMB, and OpenMM) have relatively permissive licenses like MIT license and support non-commercial academic use.

(4) Installation
> * The installation process is quite complex. As a user, I'd have to register and download NUPACK and MMB, as well as edit a series of files in order to get a functional installation. This is simply a suggestion for the developers to keep in mind.
- Our response: I agree with the reviewer on the hurdles of installing NUPACK and MMB. We thought about encapsulating everything into a python package that a user can simply import but did not proceed that way because we would like our simulation pipeline to be open for user customization including allowing a user to change the software package used in each module.
> * Related to the point above, have the authors considered using conda directly to install their software, instead of a custom shell script? `pdbfixer` is available as a conda package, and you could specify pip packages there too, e.g. `lightdock`. The installation could be reduced to a simple: 1) install nupack 2) install mmb 3) run `conda env create -f e2edna-env.yml`.
- Our response: Yes, we used conda to install the software in the JOSS branch, as shown in the "macos_installation.sh". We basically carried out the same three steps as the reviewer suggested in a slightly different order. The bash script can be found [here](https://github.com/siminegroup/E2EDNA2/blob/JOSS/macos_installation.sh).
- Change to the code: `pdbfixer` and `lightdock` packages have been added to the `e2edna-env.yml` file.
> * On this last point, the authors should strip the granular version of the env `yaml` file otherwise conda will struggle with versions on anything but the authors' hardware.
- Our response: Thanks for the suggestion. We have updated the `e2edna-env.yml` by removing all the "build specification" strings. The “e2edna-env.yml” now only contains names and versions of packages and their dependencies.
> * According to the README, the code is only tested on MacOS, although I'd imagine the most use would be on a compute cluster running Linux. Have the authors tried running their code on Linux?
- Our response: Yes, we have had this code installed and running on a Linux cluster. We provided installation instructions and automated tests on macOS only for review purposes but there are no inherent limitations on the OS.

(5) Misc
> * In several sections of their documentation, the authors mention "OpenDNA". Was this the previous name of this package?
- Our response: Yes, “OpenDNA” was an older name before we settled down on “E2EDNA 2.0”. Also, the files which mentioned “OpenDNA” were not necessary for this submission therefore were deleted in the JOSS branch.
> * It would be greatly beneficial for a user to have config files with installation paths, simulation settings etc, instead of having to edit source code. Would the authors be open to this change?
- Our response: Thanks for the suggestion. Yes, we are open to set up such config files (such as in .yml format).

(6) Comments on the Manuscript
> * In the "Statement of Need", the authors mention an "all-python" package several times. Being pedantic, this is not entirely true as their code relies on quite some compiled code in their dependencies (lightdock, openmm).
- Our response: Thanks for pointing it out. We agree that some modules are not fully written in python but the code is easy to install and interface with other python modules. For this reason we have rephrased “all-python” to "python-interfacing".
- Changes to the manuscript: Updated two places where “all-python” had been used: “all-python simulation package” and “all-python implementation”
