### 2022-03-27 a8803ca
nupack dangle stacking
Added and tested the argument `stacking` when constructing a nupack object. The argument is to control if we consider coaxial (between paired bases) stacking and dangle stacking (between an unpaired base and a base pair) in free energy calculation, therefore affecting the predicted secondary structure.
