+++ nupack-result.txt
@@ -1,0 +1,3 @@
+Complex results:
+  Complex      Pfunc dG (kcal/mol) MFE (kcal/mol)
+0      AA  8.9619e+0        -1.299         -0.350
