+++ old/afile.txt
@@ -1,221 +1,0 @@
-Comparing files amoeba2013.xml and AMOEBA-BIO-2018.XML
-***** amoeba2013.xml
- <Info>
-  <Source>amoebapro13.prm</Source>
-  <DateGenerated>2015-02-19</DateGenerated>
-  <Reference>Yue Shi, Zhen Xia, Jiajing Zhang, Robert Best, Chuanjie Wu, Jay W. Ponder, and Pengyu Ren. Polarizable Atomic Mult
-ipole-Based AMOEBA Force Field for Proteins. Journal of Chemical Theory and Computation, 9(9):4046–4063, 2013.</Reference>
- </Info>
-***** AMOEBA-BIO-2018.XML
- <Info>
-  <Source>amoebabio18.prm</Source>
-  <DateGenerated>2021-01-28</DateGenerated>
-  <Reference></Reference>
- </Info>
-*****
-
-Resync Failed.  Files are too different.
-***** amoeba2013.xml
-  <Type name="246" class="6" element="H" mass="1.008"/>
-  <Type name="247" class="43" element="O" mass="15.999"/>
-  <Type name="248" class="44" element="H" mass="1.008"/>
-  <Type name="249" class="45" element="LI" mass="6.941"/>
-  <Type name="250" class="46" element="NA" mass="22.990"/>
-  <Type name="251" class="47" element="K" mass="39.098"/>
-  <Type name="252" class="48" element="RB" mass="85.468"/>
-  <Type name="253" class="49" element="CS" mass="132.905"/>
-  <Type name="254" class="50" element="BE" mass="9.012"/>
-  <Type name="255" class="51" element="MG" mass="24.305"/>
-  <Type name="256" class="52" element="CA" mass="40.078"/>
-  <Type name="257" class="53" element="ZN" mass="65.390"/>
-  <Type name="258" class="54" element="Cl" mass="35.453"/>
- </AtomTypes>
- <Residues>
-  <Residue name="ALA">
-   <Atom name="C" type="9" />
-   <Atom name="CA" type="8" />
-   <Atom name="CB" type="13" />
-   <Atom name="H" type="10" />
-   <Atom name="HA" type="12" />
-   <Atom name="HB1" type="14" />
-   <Atom name="HB2" type="14" />
-   <Atom name="HB3" type="14" />
-   <Atom name="N" type="7" />
-   <Atom name="O" type="11" />
-   <Bond from="0" to="1" />
-   <Bond from="0" to="9" />
-   <Bond from="1" to="2" />
-   <Bond from="1" to="4" />
-   <Bond from="1" to="8" />
-   <Bond from="2" to="5" />
-   <Bond from="2" to="7" />
-   <Bond from="2" to="6" />
-   <Bond from="3" to="8" />
-   <ExternalBond from="8" />
-   <ExternalBond from="0" />
- </Residue>
-  <Residue name="ARG">
-   <Atom name="C" type="9" />
-   <Atom name="CA" type="8" />
-   <Atom name="CB" type="202" />
-   <Atom name="CD" type="206" />
-   <Atom name="CG" type="204" />
-   <Atom name="CZ" type="210" />
-   <Atom name="H" type="10" />
-   <Atom name="HA" type="12" />
-   <Atom name="HB2" type="203" />
-   <Atom name="HB3" type="203" />
-   <Atom name="HD2" type="207" />
-   <Atom name="HD3" type="207" />
-   <Atom name="HE" type="209" />
-   <Atom name="HG2" type="205" />
-   <Atom name="HG3" type="205" />
-   <Atom name="HH11" type="212" />
-   <Atom name="HH12" type="212" />
-   <Atom name="HH21" type="212" />
[…]
