+++ old/testoutput3.txt
@@ -1,1506 +1,0 @@
-HETATM    1 PD1  MOL B   1       0.000   0.000   0.000                      Pd
-HETATM    2 PD1  MOL B   2       0.000   1.945   1.945                      Pd
-HETATM    3 PD1  MOL B   3       1.945   0.000   1.945                      Pd
-HETATM    4 PD1  MOL B   4       1.945   1.945   0.000                      Pd
-HETATM    5 PD1  MOL B   5      -3.891  -9.727 -13.617                      Pd
-HETATM    6 PD1  MOL B   6       0.000  -9.727 -13.617                      Pd
-HETATM    7 PD1  MOL B   7       3.891  -9.727 -13.617                      Pd
-HETATM    8 PD1  MOL B   8      -7.781  -5.836 -13.617                      Pd
-HETATM    9 PD1  MOL B   9      -5.836  -7.781 -13.617                      Pd
-HETATM   10 PD1  MOL B  10      -3.891  -5.836 -13.617                      Pd
-HETATM   11 PD1  MOL B  11      -1.945  -7.781 -13.617                      Pd
-HETATM   12 PD1  MOL B  12      -1.945  -5.836 -15.563                      Pd
-HETATM   13 PD1  MOL B  13       0.000  -7.781 -15.563                      Pd
-HETATM   14 PD1  MOL B  14       0.000  -5.836 -13.617                      Pd
-HETATM   15 PD1  MOL B  15       1.945  -7.781 -13.617                      Pd
-HETATM   16 PD1  MOL B  16       1.945  -5.836 -15.563                      Pd
-HETATM   17 PD1  MOL B  17       3.891  -5.836 -13.617                      Pd
-HETATM   18 PD1  MOL B  18       5.836  -7.781 -13.617                      Pd
-HETATM   19 PD1  MOL B  19       7.781  -5.836 -13.617                      Pd
-HETATM   20 PD1  MOL B  20      -9.727  -3.891 -13.617                      Pd
-HETATM   21 PD1  MOL B  21      -7.781  -1.945 -13.617                      Pd
-HETATM   22 PD1  MOL B  22      -5.836  -3.891 -13.617                      Pd
-HETATM   23 PD1  MOL B  23      -5.836  -1.945 -15.563                      Pd
-HETATM   24 PD1  MOL B  24      -3.891  -3.891 -15.563                      Pd
-HETATM   25 PD1  MOL B  25      -3.891  -1.945 -13.617                      Pd
-HETATM   26 PD1  MOL B  26      -1.945  -3.891 -13.617                      Pd
-HETATM   27 PD1  MOL B  27      -1.945  -1.945 -15.563                      Pd
-HETATM   28 PD1  MOL B  28       0.000  -3.891 -15.563                      Pd
-HETATM   29 PD1  MOL B  29       0.000  -1.945 -13.617                      Pd
-HETATM   30 PD1  MOL B  30       1.945  -3.891 -13.617                      Pd
-HETATM   31 PD1  MOL B  31       1.945  -1.945 -15.563                      Pd
-HETATM   32 PD1  MOL B  32       3.891  -3.891 -15.563                      Pd
-HETATM   33 PD1  MOL B  33       3.891  -1.945 -13.617                      Pd
-HETATM   34 PD1  MOL B  34       5.836  -3.891 -13.617                      Pd
-HETATM   35 PD1  MOL B  35       5.836  -1.945 -15.563                      Pd
-HETATM   36 PD1  MOL B  36       7.781  -1.945 -13.617                      Pd
-HETATM   37 PD1  MOL B  37       9.727  -3.891 -13.617                      Pd
-HETATM   38 PD1  MOL B  38      -9.727   0.000 -13.617                      Pd
[…]
