### 2022-03-27 f2178e2
Delete CONTRIBUTING.md
