+++ CONTRIBUTING.md
@@ -1,15 +1,0 @@
-## How to Contribute to E2EDNA Development
-E2EDNA is an ongoing project. We welcome anyone who wants to contribute, whether by adding a feature, fixing a bug, improving documentation, and etc. 
-The process is quite simple, as follows.
-
-Option1: start with an issue report to describe any issue or suggestion. Then we can discuss significance, solutions, designs, and etc.
-
-Option2: directly contact us ([The Simine Group](https://www.siminegroup.ca/)).
-
-The next step is to [create a pull request](https://help.github.com/en/articles/about-pull-requests) with the code changes. 
-If necessary, put the "WIP" in the title of the pull request to indicate it is still work in progress. 
-Once the pull request is reviewed, tested and approved, it will be merged. 
-Congratulations on a succesful contribution!
-
-Of course, you are more than welcome to fork this project, develop it on your own, then eventually contribute back to this repository. 
-We believe that it will be beneficial to all users if the main development activities are centered around this repository.
