+++ README.MD
@@ -1,129 +1,0 @@
-# Documentation
-## E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
-
-### An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
-
-**Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine**
-
-* For original version of E2EDNA: J. Chem. Inf. Model. 2021, 61, 9, 4139–4144 (https://doi.org/10.1021/acs.jcim.1c00696) and https://github.com/InfluenceFunctional/E2EDNA
-
-* Interested in contributing to developing E2EDNA? Check out how to contribute [here](https://github.com/siminegroup/E2EDNA2/blob/main/CONTRIBUTING.md).
-## 1. Installation
-This installation path has been tested on macOS and it relies on conda and pip package managers. 
-1. Download the E2EDNA 2.0 package from this repository.
-2. Register and download NUPACK package from http://www.nupack.org/downloads, find out the path to the downloaded package, such as ``~/Downloads/nupack-4.0.0.27/package``.
-3. In the E2EDNA2 directory please modify the [macos_installation.sh](https://github.com/siminegroup/E2EDNA2/blob/main/macos_installation.sh) script: update the path to NUPACK installation package (see step 2).
-4. From the E2EDNA2 directory run ``source macos_installation.sh`` in command line to install a virtual environment "e2edna" and required dependences. Note that (e2edna) environment should be activated when the script finishes. 
-   * Caveat: If the scripts fails to activate the environment automatically, this is likely because `conda activate e2edna` command gives an error such as "Your shell has not been properly configured to use 'conda activate'".
-   * To fix this issue, please, first remove the installed environment by ``conda remove --name e2edna --all`` in command line, then either replace the conda activation command in the script with `source activate /path-to-envs/e2edna`, e.g., `source activate ~/miniconda3/envs/e2edna` and run the script again
-   * OR run the commands one by one interactively in your terminal. 
-5. Register and download MMB from https://simtk.org/projects/rnatoolbox. Place the "Installer.*version.OS*" folder, such as "Installer.3_0.OSX" into the E2EDNA2 directory. N.B.: Do not specify `DYLD_LIBRARY_PATH` for macOS against the recommendations of the MMB installation guide. This is to avoid interference with the OpenMM module.
-6. Update 3 paths in [main.py](https://github.com/siminegroup/E2EDNA2/blob/main/main.py):
-```
-params['workdir'] = '/path-to-E2EDNA2-main/localruns'                                # directory manually created to store all future jobs
-params['mmb dir'] = '/path-to-E2EDNA2-main/Installer.version.OS/lib'                 # path to MMB dylib files
-params['mmb']     = '/path-to-E2EDNA2-main/Installer.version.OS/bin/MMB-executable'  # path to MMB executable
-```
-## 2. Running a job
-### Quickstart
-* Set 3 paths in `params`, as indicated in step6 of [Installation](https://github.com/siminegroup/E2EDNA2#1-installation) section. Leave the other keys of `params` unchanged.
-* Run the bash script [automate_tests.sh](https://github.com/siminegroup/E2EDNA2/blob/main/automate_tests.sh) to test run all 8 modes automatically. Check out [Test runs](https://github.com/siminegroup/E2EDNA2#4-test-runs-inputs-and-outcomes) section for details. 
[…]
