### 2022-03-27 f3b2570
Delete automate_tests.sh
