+++ automate_tests.sh
@@ -1,77 +1,0 @@
-#!/bin/bash
-
-# Assumption: lcoal, macos; user has defined work dir, MMB path in main.py and provided target ligand's pdb etc.
-# in short, it is ready to run and test different modes
-
-printf "Start automating tests one by one...\n"
-printf "===================================="
-
-printf "\nTESTING MODE #1: '2d structure'\n"
-python main.py --run_num=1 --mode='2d structure' --aptamerSeq='TAATGTTAATTG' --ligand='False' --ligandType='' --ligandSeq='' #> run1.out
-printf "\nEND OF TEST #1. Results are saved to folder \"run1\", where:\n"
-printf "\t2d structure: in record.txt\n"
-printf "===================================="
-
-printf "\nTESTING MODE #2: '3d coarse'\n"
-python main.py --run_num=2 --mode='3d coarse' --aptamerSeq='TAATGTTAATTG' --ligand='False' --ligandType='' --ligandSeq='' #> run2.out
-printf "\nEND OF TEST #2. Results are saved to folder \"run2\", where:\n"
-printf "\t2d structure: in record.txt\n"
-printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
-printf "===================================="
-
-printf "\nTESTING MODE #3: '3d smooth'\n"
-python main.py --run_num=3 --mode='3d smooth' --aptamerSeq='TAATGTTAATTG' --ligand='False' --ligandType='' --ligandSeq='' #> run3.out
-printf "\nEND OF TEST #3. Results are saved to folder \"run3\", where:\n"
-printf "\t2d structure: in record.txt\n"
-printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
-printf "\tShort MD relaxation trajectory of free aptamer: foldedAptamer_0_processed_trajectory.dcd and clean_foldedAptamer_0_processed_trajectory.dcd (without solvent and ions)\n"
-printf "\tRelaxed aptamer structure: relaxedAptamer_0.pdb\n"
-printf "===================================="
-
-printf "\nTESTING MODE #4: 'coarse dock'\n"
-python main.py --run_num=4 --mode='coarse dock' --aptamerSeq='TAATGTTAATTG' --ligand='YQTQ.pdb' --ligandType='peptide' --ligandSeq='YQTQTNSPRRAR' #> run4.out
-printf "\nEND OF TEST #4. Results are saved to folder \"run4\", where:\n"
-printf "\t2d structure: in record.txt\n"
-printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
-printf "\tMMB-folded aptamer docked by target ligand: complex_0_0.pdb (if docking happened)\n"
-printf "===================================="
-
-printf "\nTESTING MODE #5: 'smooth dock'\n"
-python main.py --run_num=5 --mode='smooth dock' --aptamerSeq='TAATGTTAATTG' --ligand='YQTQ.pdb' --ligandType='peptide' --ligandSeq='YQTQTNSPRRAR' #> run5.out
-printf "\nEND OF TEST #5. Results are saved to folder \"run5\", where:\n"
-printf "\t2d structure: in record.txt\n"
-printf "\tMMB-folded aptamer structure: foldedAptamer_0.pdb\n"
-printf "\tShort MD relaxation trajectory of free aptamer: foldedAptamer_0_processed_trajectory.dcd and clean_foldedAptamer_0_processed_trajectory.dcd (without solvent and ions)\n"
-printf "\tRelaxed aptamer structure: relaxedAptamer_0.pdb\n"
-printf "\tRelaxed aptamer docked by target ligand: complex_0_0.pdb (if docking happened)\n"
[…]
