+++ e2edna-env.yml
@@ -1,173 +1,0 @@
-name: e2edna
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - ambertools=21.11=py39hf80593e_0
-  - arpack=3.7.0=hefb7bc6_2
-  - astunparse=1.6.3=pyhd8ed1ab_0
-  - biopython=1.78=py39h9ed2024_0
-  - blas=1.0=mkl
-  - blosc=1.21.0=he49afe7_0
-  - boost-cpp=1.74.0=hf3dc895_5
-  - brotli=1.0.9=h0d85af4_6
-  - brotli-bin=1.0.9=h0d85af4_6
-  - bzip2=1.0.8=h0d85af4_4
-  - c-ares=1.18.1=h0d85af4_0
-  - c-compiler=1.3.0=h0d85af4_0
-  - ca-certificates=2021.10.8=h033912b_0
-  - cctools=973.0.1=hd9211c8_5
-  - cctools_osx-64=973.0.1=hb41342b_5
-  - certifi=2021.10.8=py39h6e9494a_1
-  - cftime=1.5.1.1=py39hc89836e_1
-  - clang=11.1.0=h694c41f_1
-  - clang-11=11.1.0=default_he082bbe_1
-  - clang_osx-64=11.1.0=hc8532ff_6
-  - clangxx=11.1.0=default_he082bbe_1
-  - clangxx_osx-64=11.1.0=h2c55bc3_6
-  - colorama=0.4.4=pyh9f0ad1d_0
-  - compiler-rt=11.1.0=h654b07c_0
-  - compiler-rt_osx-64=11.1.0=h8c5fa43_0
-  - compilers=1.3.0=h694c41f_0
-  - curl=7.81.0=hf45b732_0
-  - cxx-compiler=1.3.0=h940c156_0
-  - cycler=0.11.0=pyhd8ed1ab_0
-  - cython=0.29.26=py39h9fcab8e_0
-  - fftw=3.3.10=nompi_hf082fe4_102
-  - fonttools=4.28.5=py39h89e85a6_0
-  - fortran-compiler=1.3.0=h6818c3e_0
-  - freetype=2.10.4=h4cff582_1
-  - gfortran=9.3.0=h768ea0c_15
-  - gfortran_impl_osx-64=9.3.0=h9cc0e5e_23
-  - gfortran_osx-64=9.3.0=h18f7dce_15
-  - gmp=6.2.1=h2e338ed_0
-  - griddataformats=0.6.0=pyhd8ed1ab_0
-  - gsd=2.5.1=py39hc89836e_0
-  - hdf4=4.2.15=hefd3b78_3
-  - hdf5=1.12.1=nompi_h2f0ef1a_103
-  - icu=69.1=he49afe7_0
-  - intel-openmp=2021.4.0=hecd8cb5_3538
-  - isl=0.22.1=hb1e8313_2
-  - jbig=2.1=h0d85af4_2003
-  - joblib=1.1.0=pyhd8ed1ab_0
-  - jpeg=9d=hbcb3906_0
-  - khronos-opencl-icd-loader=2022.01.04=h0d85af4_0
-  - kiwisolver=1.3.2=py39hf018cea_1
-  - krb5=1.19.2=hcfbf3a7_3
-  - lcms2=2.12=h577c468_0
-  - ld64=609=hd2e7500_5
-  - ld64_osx-64=609=h6adca69_5
-  - lerc=3.0=he49afe7_0
-  - libblas=3.9.0=12_osx64_mkl
-  - libbrotlicommon=1.0.9=h0d85af4_6
[…]
