+++ leap_template.in
@@ -1,6 +1,0 @@
-# TARGET_IS_PEPTIDE source leaprc.protein.ff14SB
-source leaprc.DNA_FF  # leaprc.DNA.OL15 or leaprc.DNA.bsc1
-# TARGET_IS_RNA source leaprc.RNA.OL3
-myPDB = loadPdb "mySTRUCTURE_amb_processed.pdb"
-saveAmberParm myPDB mySTRUCTURE_amb_processed.top mySTRUCTURE_amb_processed.crd
-quit
