### 2022-03-27 75b157f
Delete macos_installation.sh
