### 2022-03-27 e8b39a1
Delete main.py
