+++ opendna.py
@@ -1,767 +1,0 @@
-import sys
-import glob
-from shutil import copyfile, copytree
-
-import interfaces
-from utils import *
-from analysisTools import *
-
-# noinspection PyPep8Naming  # ignore inspetion based on python PEP8 Naming convention
-
-
-class opendna:
-    def __init__(self, params):
-        self.workDir = ""  # rather use an empty string, not an empty list
-        self.params = params
-        self.aptamerSeq = self.params['aptamerSeq']        
-        self.targetPDB = self.params['target ligand']        # either pdb filename (eg, "peptide.pdb") or boolean ``False``
-        self.targetType = self.params['target ligand type']  # 'peptide' or 'DNA' or 'RNA' or 'other'. Assume now the target is a peptide
-        self.targetSeq = self.params['target sequence']
-        # empty string, if target is not a peptide/DNA/RNA: how to deal with the "analyzeBinding"? No analysis in this simulator.
-        
-        self.pdbDict = {}
-        self.dcdDict = {}
-        self.actionDict = {}
-        self.getActionDict()  # specify the actions based on the selected mode
-        # modify actions based on three toggles.
-        if self.params['skip MMB'] is True:  
-            self.actionDict['do 2d analysis'] = False  
-            self.actionDict['do MMB'] = False        
-        if self.params['pick up from freeAptamerChk'] is True:
-            # everything before "freeAptamerDynamics" is skipped
-            self.actionDict['do 2d analysis'] = False
-            self.actionDict['do MMB'] = False            
-            self.actionDict['do smoothing'] = False
-            self.actionDict['get equil repStructure'] = True  # making sure to run "freeAptamerDynamics"
-            # # Not used for now:
-            # printRecord('Copying the checkpoint file: {} to workdir.'.format(self.params['chk file']), self.workDir+'/')
-            # if os.path.exists(self.params['chk file']):
-            #     copyfile(self.params['chk file'], self.workDir + '/' + self.params['chk file'])
-            # else:
-            #     printRecord('Designated checkpoint file does not exist! Terminating the pipeline.', self.workDir+'/')
-            #     self.terminateRun()
-        elif self.params['pick up from complexChk'] is True:
-            # everything before "complexDynamics" is skipped
-            self.actionDict['do 2d analysis'] = False
-            self.actionDict['do MMB'] = False            
-            self.actionDict['do smoothing'] = False            
-            self.actionDict['get equil repStructure'] = False
-            self.actionDict['do docking'] = False
-            self.actionDict['do binding'] = True  # making sure to run "complexDynamics"
-
-        if self.actionDict['make workdir']:
-            self.setup()  # if we don't need a workdir & MMB files (eg, give a 3D structure), don't make one.
-
-        self.i = int(-1)
-        self.j = int(-1)
-
-    def getActionDict(self):
-        """
-        generate a binary sequence of 'to-do's', given user's input into self.actionDict
-        :return: none
-        """
-        if self.params['mode'] == '2d structure':
-            self.actionDict['make workdir'] = True
-            self.actionDict['do 2d analysis'] = True
-            self.actionDict['do MMB'] = False
-            self.actionDict['do smoothing'] = False
-            self.actionDict['get equil repStructure'] = False
-            self.actionDict['do docking'] = False
-            self.actionDict['do binding'] = False
-
-        elif self.params['mode'] == '3d coarse':
-            self.actionDict['make workdir'] = True
-            self.actionDict['do 2d analysis'] = True
-            self.actionDict['do MMB'] = True
-            self.actionDict['do smoothing'] = False
-            self.actionDict['get equil repStructure'] = False
-            self.actionDict['do docking'] = False
-            self.actionDict['do binding'] = False
-
[…]
