+++ old/processTinkerForcefield.txt
@@ -1,1162 +1,0 @@
-# =============================================================================================
-# MODULE DOCSTRING
-# =============================================================================================
-
-"""
-processTinkerForceField.py
-
-   Convert TINKER force field files into xml files for use by pyopenmm
-
-   (1) read residue template file
-   (2) read TINKER parameter file
-   (3) assign biotypes to each atom in residue template file
-   (4) output force-field parameter file
-
-"""
-# =============================================================================================
-# GLOBAL IMPORTS
-# =============================================================================================
-
-import os
-import xml.etree.ElementTree as etree
-import sys
-import shlex
-import math
-import datetime
-import os.path
-
-# =============================================================================================
-# Ion list
-# =============================================================================================
-
-# biotype    2003    NA      "Sodium Ion"                      250
-# biotype    2004    K       "Potassium Ion"                   251
-# biotype    2005    MG      "Magnesium Ion"                   255
-# biotype    2006    CA      "Calcium Ion"                     256
-# biotype    2007    CL      "Chloride Ion"                    258
-
-ions = {'Li+': ['LI', 249],
-        'Na+': ['NA', 250],
-        'K+': ['K', 251],
-        'Rb+': ['RB', 252],
-        'Cs+': ['CS', 253],
-        'Be+': ['BE', 254],
-        'Mg+': ['MG', 255],
-        'Ca+': ['CA', 256],
-        'Zn+': ['ZN', 257],
-        'Cl-': ['Cl', 258]
-        }
-
-atomTypes = {}
-bioTypes = {}
-
-
-# =============================================================================================
-# Default 'constructor' for atoms
-# =============================================================================================
-
-def getDefaultAtom():
-    atom = dict();
-    atom['tinkerLookupName'] = 'XXX'
-    atom['type'] = -1
-    atom['bonds'] = dict()
-
-    return atom
-
-
-# =============================================================================================
-# Add bond to atomDict[]; atoms are added to atomDict[] if missing
-# =============================================================================================
-
-def addBond(atomDict, atom1, atom2):
-    if (atom1 not in atomDict):
-        atomDict[atom1] = getDefaultAtom()
-
-    if (atom2 not in atomDict):
-        atomDict[atom2] = getDefaultAtom()
-
-    atomDict[atom2]['bonds'][atom1] = 1
-    atomDict[atom1]['bonds'][atom2] = 1
-
-
-# =============================================================================================
-# Get atom dictionary from xml atom list
-# =============================================================================================
-
-def getXmlAtoms(atoms):
-    atomInfo = dict();
-    for atom in atoms:
-        name = atom.attrib['name']
-        atomInfo[name] = getDefaultAtom()
-        atomInfo[name]['tinkerLookupName'] = atom.attrib['tinkerLookupName']
-
-    return atomInfo
-
-
-# =============================================================================================
-# Get bond dictionary from xml bond list
-# =============================================================================================
-
-def getXmlBonds(bonds):
-    bondInfo = dict();
-    for bond in bonds:
-        atom1 = bond.attrib['from']
-        atom2 = bond.attrib['to']
-        if (atom1 not in bondInfo):
-            bondInfo[atom1] = dict()
-        if (atom2 not in bondInfo):
-            bondInfo[atom2] = dict()
-
-        bondInfo[atom1][atom2] = 1
-        bondInfo[atom2][atom1] = 1
-
-    return bondInfo
-
-
[…]
