+++ README_EXAMPLE.md
@@ -1,17 +1,0 @@
-# Example
-
-We will demonstrate the pipeline on a `full binding` simulation of a simple hairpin aptamer with a short peptide. 
-
-`sequence='CGCTTTGCG'`
-`peptide='YYYYYY'`
-
-## Parameters
-
-
-## Results
-
-#### 2D Fold
-#### 3D Coarse Fold
-#### 3D Representative Structure
-#### Docking Simulation
-#### Binding Simulation
