+++ README.MD
@@ -1,141 +1,0 @@
-# E2EDNA - OpenMM Implementation - OpenDNA!
-# This branch is created for running the pipeline on computing clusters such as Compute Canada.
-
-### A fully-automated code for DNA aptamer 2D and 3D structure analysis, and binding analysis with small molecules and short peptides.
-
-**Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine**
-
-_mjakilgour gmail com_
-
-## INSTALLATION
-* Get the repository
-```
-git clone git@github.com:InfluenceFunctional/OpenDNA.git WHAT_YOU_WANT_IT_TO_BE_CALLED
-```
-* Setup python environment with the following packages
-  * Note to Simine Group Members running on Compute Canada - skip this section and go directly to paths. Your virtual environment is automatically setup via the instructions in EXAMPLE_SUB.sh
-```
-biopython==1.78
-certifi==2020.12.5
-cycler==0.10.0
-decorator==4.4.2
-GridDataFormats==0.5.0
-gsd==1.9.3
-joblib==1.0.1
-kiwisolver==1.3.1
-matplotlib==3.3.2
-mmtf-python==1.1.2
-mock==4.0.3
-msgpack==0.6.2
-networkx==2.5
-pandas==1.1.3
-ParmEd==3.2.0
-Pillow==7.2.0
-pyparsing==2.4.7
-python-dateutil==2.8.1
-pytz==2021.1
-scipy==1.4.1
-six==1.15.0
-tqdm==4.59.0
-scikit_learn==0.24.2
-Cython==0.29.23
-mdtraj==1.9.5
-```
-* Download appropriate MacroMoleculeBuilder for your system from [SimTK here](https://simtk.org/projects/rnatoolbox) - old and new versions have been tested and seem to work. We'll need the 'Installer' directory and MMB executable. 
-* Create working directory where runs will take place on your system. 
-  Each new run will create a serially-numbered directory in this folder containing all the relevant outputs.
-* Set paths in main.py and submission shell script, if relevant (example enclosed). Note, there is a toggle for running on a 'local' machine vs 'cluster', with distinct paths for e.g., developing on a local vs cluster environment.
-```
-* params['mmb'] --> set to the MacroMoleculeBuilder executable on your system
-* params['mmb dir'] --> set to the MMB directory, usually a name like /Installer.#_##.Linux
-- params['mmb template'] and params['mmb params'] --> included in this repository, should not have to be changed
-* params['ld path'] --> paths to various lightdock python scripts (included in this repository, alternatively automatically come with a lightdock pip installation). Depending on your installation, you may need to call 'python' before these scripts in the path.
-* params['workdir'] --> the working directory you created in the previous step
-```
-
-## Modes
-OpenDNA takes in a DNA aptamer sequence in FASTA format, and optionally a short peptide or other small molecule, and returns details of the aptamer structure and binding behaviour.
-This code implements several distinct analysis modes so users may customize the level of computational cost and accuracy.
-
-* `2d structure` &rarr; returns NUPACK or seqfold analysis of aptamer secondary structure. Very fast, O(<1s). If using NUPACK, includes probability of observing a certain fold and of suboptimal folds within kT of the minimum.
-* `3d coarse` &rarr; returns MMB fold of the best secondary structure. Fast O(5-30 mins). Results in a strained 3D structure which obeys base pairing rules and certain stacking interactions.
-* `3d smooth` &rarr; identical to '3d coarse', with a short MD relaxation in solvent. ~Less than double the cost of '3d coarse' depending on relaxation time.
[…]
