### 2022-03-29 d6bf52d
Update interfaces.py: nupack and lightdock classes
