+++ interfaces.py
@@ -1,6 +1,6 @@
 """
-interfaces for other major software used by OpenDNA
-MacroMolecule Builder (MMB): installed PeptideBuilder on pip. Is it enough?
+interfaces for other major software used by E2EDNA
+MacroMolecule Builder (MMB)
 OpenMM (omm)
 LightDock (ld)
 NUPACK
@@ -13,20 +13,16 @@ from nupack import *
 from simtk.openmm import *  # In fact what happens under the hood: from openmm import *
 from simtk.openmm.app import *
 import simtk.unit as unit
-# from openmm import *
-# from openmm.app import *
-# import openmm.unit as unit
-
-from utils import *
-from analysisTools import *
-
 # installed openmm: http://docs.openmm.org/7.5.0/userguide/application.html#installing-openmm
 # openmm is installed as an individual package;
 # package simtk is created with packages "openmm" and "unit" inside but literally import the stand-alone openmm.
 
+from utils import *
+from analysisTools import *
+
 
 class nupack:
-    def __init__(self, sequence, temperature, ionicStrength, mgConc=0):
+    def __init__(self, aptamerSeq, temperature, ionicStrength, mgConc=0, stacking='nostacking'):
         self.sequence = sequence
         self.temperature = temperature
         self.ionicStrength = ionicStrength
@@ -40,26 +36,59 @@ class nupack:
 
     def energyCalc(self):
         """
-        sequence is DNA FASTA format
+        aptamerSeq is in FASTA format
         temperature in C or K
         ionicStrength in Molar
         Ouput a lot of analysis results in self.output
         :return:
         """
-        if self.temperature > 273:  # auto-detect Kelvins
+        # Compute free energy gap used by looking for suboptimal structure whose energy <= delta_G_mfe + gap
+        if self.temperature > 273.15:  # auto-detect Kelvins
             gap = 2 * self.R * self.temperature
-            CelsiusTemprature = self.temperature - 273
+            CelsiusTemprature = self.temperature - 273.15
         else:
-            gap = 2 * self.R * (self.temperature + 273)  # convert to Kelvin fir kT
+            gap = 2 * self.R * (self.temperature + 273.15)  # convert to Kelvin for kT
             CelsiusTemprature = self.temperature
 
-        A = Strand(self.sequence, name='A')
-        comp = Complex([A], name='AA')
-        set1 = ComplexSet(strands=[A], complexes=SetSpec(max_size=1, include=[comp]))
-        model1 = Model(material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
-        results = complex_analysis(set1, model=model1, compute=['pfunc', 'mfe', 'subopt', 'pairs'], options={'energy_gap': gap})
-        self.output = results[comp]
-    # TODO: what do these nupack functions do?
+        A = Strand(self.aptamerSeq, name='A') # specify a strand and name
+        # A.nt() # calculate the number of nucleotides
+        comp = Complex([A], name='AA', bonus=0) # specify a complex of one or more interacting strands and name is optional.
+        
+        # specify a set of strands that interact to form a set of complexes
+        set1 = ComplexSet(strands=[A], complexes=SetSpec(max_size=1, include=[comp])) # 1 complex set: [[A]]
+
+        model1 = Model(ensemble=self.stacking, material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
+        '''
+        - ensemble: whether to consider coaxial or dangle stacking for multiloop or exterior loop. 
+            - 'stacking': default. Complex ensemble with coaxial and dangle stacking
[…]
