+++ main.py
@@ -1,153 +1,189 @@
-"""
-from typing import Dict, Any, Union  # for the dictionary of params
-...
-params: Dict[str, Union[Union[str, int, bool, float], Any]] = {}  # set up as a dictionary
-==> To replace the params = {}
-"""
+'''
+E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
+
+An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
+
+Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine
+
+Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+'''
 from opendna import *
-# from utils import *
-# from simtk.openmm.app import *
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
 params = {}
-params['device'] = 'cluster'  # 'local' or 'cluster'
-params['platform'] = 'CUDA'  # 'CUDA' or 'CPU'
-params['platform precision'] = 'single'  # 'single' or 'double'. Only relevant on 'CUDA' platform
-
-if params['device'] == 'cluster':
-    params['local device platform'] = ""  # leave it blank or specify the cluster's OS. This is for running MMB. Consult the admin for cluster if running into problems.
-    cmdLineInputs = get_input()  # get input arguments from command lines
-                                 # remember to edit the method "get_input()" in utils.py!!!`
-    params['run num'] = cmdLineInputs[0]   # option to get run num from command line (default zero)
-    params['sequence'] = cmdLineInputs[1]  # DNA aptamer sequence
-    params['peptide'] = cmdLineInputs[2]   # target peptide sequence
-    params['max walltime'] = cmdLineInputs[3]  # maximum walltime in hours
-                                               # After free aptamer MD, "sampling time" might be reduced to ensure bindingDynamics to finish before time out.
-                                               # Or give up if this isn't enough time to complete even a minimum run.
-    # Physical params
-    params['temperature'] = cmdLineInputs[4]     # Kevin - used to predict secondary structure and for MD thermostat
-    params['pH'] = cmdLineInputs[5]              # simulation will automatically protonate the peptide up to this pH. Used in OpenMM for waterBox
-    params['ionicStrength'] = cmdLineInputs[6]   # Molar - sodium concentration - used to predict secondary structure and add ions to simulation box, must be 1100 M > [Na] > 50 for nupack to run
-                                                 # TODO: how about adding other ions? Expand the FF as well?
-    params['[Mg]'] = cmdLineInputs[7]            # Molar - magnesium concentration: 0.2 M > [Mg] > 0 - ONLY applies to NuPack fold - Does NOT add Mg to MD simulations
-    params['impSolv'] = cmdLineInputs[8]         # this is a string, cannot be used as parameter in prmtop.createSys()!
-    params['pressure'] = 1                       # atmosphere
-
-elif params['device'] == 'local':
-    params['local device platform'] = 'macos'  # macos, or linux or windows
-    params['run num'] = 1  # manual setting, for 0, do a fresh run, for != 0, pickup on a previous run
[…]
