### 2022-03-29 68cdfbe
Delete lib/lightdock directory
