+++ pdbfixersource.py
@@ -1,1291 +1,0 @@
-"""
-pdbfixer.py: Fixes problems in PDB files
-
-This is part of the OpenMM molecular simulation toolkit originating from
-Simbios, the NIH National Center for Physics-Based Simulation of
-Biological Structures at Stanford, funded under the NIH Roadmap for
-Medical Research, grant U54 GM072970. See https://simtk.org.
-
-Portions copyright (c) 2013-2020 Stanford University and the Authors.
-Authors: Peter Eastman
-Contributors:
-
-Permission is hereby granted, free of charge, to any person obtaining a
-copy of this software and associated documentation files (the "Software"),
-to deal in the Software without restriction, including without limitation
-the rights to use, copy, modify, merge, publish, distribute, sublicense,
-and/or sell copies of the Software, and to permit persons to whom the
-Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-THE AUTHORS, CONTRIBUTORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
-OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
-USE OR OTHER DEALINGS IN THE SOFTWARE.
-"""
-from __future__ import absolute_import
-__author__ = "Peter Eastman"
-__version__ = "1.7"
-
-import simtk.openmm as mm
-import simtk.openmm.app as app
-import simtk.unit as unit
-from simtk.openmm.app.internal.pdbstructure import PdbStructure
-from simtk.openmm.app.internal.pdbx.reader.PdbxReader import PdbxReader
-from simtk.openmm.app.element import hydrogen, oxygen
-from simtk.openmm.app.forcefield import NonbondedGenerator
-
-# # # importing 'simtk.openmm' is deprecated.  Import 'openmm' instead
-# # # There is another statement down under "try" that needs update!!!!
-# import openmm as mm
-# import openmm.app as app
-# import openmm.unit as unit
-# from openmm.app.internal.pdbstructure import PdbStructure
-# from openmm.app.internal.pdbx.reader.PdbxReader import PdbxReader
-# from openmm.app.element import hydrogen, oxygen
-# from openmm.app.forcefield import NonbondedGenerator
-
-
-# Support Cythonized functions in OpenMM 7.3
-# and also implementations in older versions.
-try:
-    from simtk.openmm.app.internal import compiled
-    # from openmm.app.internal import compiled
-
-    matchResidue = compiled.matchResidueToTemplate
-except ImportError:
-    matchResidue = app.forcefield._matchResidue
-
-import numpy as np
-import numpy.linalg as lin
-import sys
-import os
-import os.path
-import math
-
-from pkg_resources import resource_filename
-
-if sys.version_info >= (3,0):
-    from urllib.request import urlopen
-    from io import StringIO
-else:
-    from urllib2 import urlopen
-    from cStringIO import StringIO
-
-substitutions = {
[…]
