### 2022-04-09 1a7fffc
Delete pdbfixersource.py
pdbfixer's source code was used initially just for development; pdbfixer has been used as a python library (installed by pip)
