### 2022-04-24 #4182 @JoaoRodrigues
Hi @csoneson, apologies, I haven't had as much time to work on this as I wanted. Replied to the authors just now. Last thing I would need before 👍 is to test the software locally. Unfortunately, the current examples (AFAIK) require some third-party software that I'd have to register to download, and I don't want to do so. I've asked the authors to provide an example that does not require those modules, since they've mentioned that that is a possibility.
