### 2022-04-24 siminegroup/E2EDNA2#14 @taoliu032
Hi @JoaoRodrigues, Thanks for your comment.

To clarify: I think that the two dependencies you referred to are NUPACK and MMB. If so, yes it is possible to bypass both of them by providing the software with a DNA structure in pdb format.

As you requested, I have made an example that does not require the dependencies of NUPACK and MMB. To do so, I needed to slightly modify the code, such as removing the import statement `from nupack import *`. So, to avoid having conflicts with the other reviewer's review process, I created another branch for the modified code, called [JOSS_test_without_requiring_nupack_and_MMB](https://github.com/siminegroup/E2EDNA2/tree/JOSS_test_without_requiring_nupack_and_MMB).

In this branch, "foldedAptamer_0.pdb" is the example DNA aptamer structure for the test. Regarding how to run the test on your local computer: for convenience, after setting up the conda environment and workdir path in `main.py`, you can simply run the last test in the `automate_tests.sh`. It will test all the modules (excluding the bypassed two) of the software. For reference, the corresponding command is also copied here: `python main.py --run_num=8 --mode='full binding' --aptamerSeq='TAATGTTAATTG' --ligand='YQTQ.pdb' --ligandType='peptide' --ligandSeq='YQTQTNSPRRAR'`
