### 2022-04-24 #4182 @taoliu032
Thank you for the comment. An example has been prepared as requested, please see the detail here: https://github.com/siminegroup/E2EDNA2/issues/14#issuecomment-1107938433.

As always, please do not hesitate to ping me should there be any question.
