### 2022-04-25 siminegroup/E2EDNA2#14 @JoaoRodrigues
In addition, here are some comments from installing and running the software:

- Running the `macos_installation.sh` script fails on my machine because I did not install `nupack`. The script has a hardcoded path for a folder. Please refactor this. I realize this is "fixed" in the branch you pointed me to, but there should be a general solution for this problem.

- You should consider having a folder called `examples` where you put the input data files needed to run particular workflows and maybe a shell script with the options. This will help users get started using your software.

- The command you wrote me before failed with an error: `FileNotFoundError: [Errno 2] No such file or directory: '/absolute-path-to-E2EDNA2-JOSS/localruns/run8'`. Please, do **not** hardcode values or paths. Use configuration files. Do not expect users to edit your source code.

- Related to the above. I had to edit the `main.py` file in a couple of lines and then create two separate folders (`scratch/` and `localruns/`) in order for the code to run. If I was a prospective user, I'd be very confused about what to do and would most likely give up. Please, consider supper user-configuration files, or at the very least separating configurations from code in a "global" yaml file for these top-level configs. Like this, will users have to copy the repository every time they want to run a new system? Or try a different set of parameters?

- The workflow should be pretty clear about what files are temporary/intermediate and what files the user should look at at the very end. There is a `record.txt` file that I assume is a log file, and should be renamed as such for clarity. Other than that, I cannot understand, from the log file and the output, which files I should look once the simulations finish. You could either write that explicitly in the log file, separate the different stages in sub-folders in the run folder, or simply rename the files "stage_0", "stage_1", ..., "final_output" or something like that.
