### 2022-04-25 #4182 @rvhonorato
@JoaoRodrigues did most of the heavy lifting addressing most of the points I had with the code, thanks ;) For me before the final OK I also just need to test it locally, going over it now based on the link by @taoliu032 above.
