### 2022-04-25 siminegroup/E2EDNA2#14 @rvhonorato
> This type of dependencies is usually implemented by isolating the part of the code requiring the third-party code in question in a function or class, and then having the import there in a try/except statement

+1

> The command you wrote me before failed with an error: FileNotFoundError: [Errno 2] No such file or directory: '/absolute-path-to-E2EDNA2-JOSS/localruns/run8'. Please, do not hardcode values or paths. Use configuration files. Do not expect users to edit your source code.

I have the same error, and agree that you should not use hardcoded values.

I see that most of the logic revolves around this `params` dictionary, a very simple solution would be to write a `configuration.json` with all the "hardcoded" values, load this during execution and add the `cmdLineInputs` as needed.
