### 2022-05-03 006af1f
Complete PR#8
Complete the changes mentioned in the comments of PR#8: https://github.com/siminegroup/E2EDNA2/pull/8
