### 2022-05-03 6c77aba
Two more fixes in main.py
1. Add an argparse argument to specify OS in command line.
2. Only overwrite the output directory (eg, ".../run1") when `-f/--force` and '--pick_up  No' in command line.
Then update the help message of main.py in the README.MD according to the new main.py
