+++ main.py
@@ -5,9 +5,7 @@ E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
 
 An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
 
-Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine
-
-Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
+Copyright 2022 Michael Kilgour, Tao Liu and Lena Simine
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -23,217 +21,120 @@ Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
 '''
 import argparse
 import glob
-from opendna import *
 import os
 import platform
 import shutil
+from opendna import *
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
 
-def get_args():
-    ''' Parse commandline arguments '''
 
-    parser = argparse.ArgumentParser(
-        description='E2EDNA: Simulate DNA aptamers complexed with target ligands',
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+parser = argparse.ArgumentParser(
+    description='E2EDNA: Simulate DNA aptamers complexed with target ligands',
+    formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 
-    run_info = parser.add_argument_group('Run Parameters')
-    system_info = parser.add_argument_group('System Configuration')
-    paths = parser.add_argument_group('Directory Settings')
+compute_info = parser.add_argument_group('Compute Platform Configuration')
+paths = parser.add_argument_group('Directory Settings')
+run_info = parser.add_argument_group('Run Parameters')
 
-    parser.add_argument('-f',
-                             '--force',
-                             action='store_true',
-                             help='Overwrite existing --run_num')
 
-    run_info.add_argument('-r',
-                             '--run_num',
-                             metavar='INT',
-                             type=int,
-                             required=True,
-                             help='Run number. Output will be written to {--workdir}/run{--run_num}/')
-    run_info.add_argument('-m',
-                             '--mode',
-                             metavar='MODE',
-                             type=str,
-                             required=True,
-                             help='Run mode',
-                             choices=['2d structure', '3d coarse', '3d smooth',
-                                      'coarse dock', 'smooth dock', 'free aptamer',
-                                      'full dock', 'full binding'])
-    run_info.add_argument('-a',
-                             '--aptamerSeq',
-                             metavar='SEQ',
-                             type=str,
-                             required=True,
-                             help="DNA Aptamer sequence (5'-3')")
-    run_info.add_argument('-l',
-                             '--ligand',
-                             metavar='PDB',
-                             type=str,
-                             default='False',
-                             required=True,
-                             help='Filename of target ligand')
-    run_info.add_argument('-t',
-                             '--ligand_type',
-                             metavar='TYPE',
-                             type=str,
-                             default='',
-                             help='Target ligand molecule type',
-                             choices=['peptide', 'DNA', 'RNA', 'other', ''])
-    run_info.add_argument('-s',
-                             '--ligand_seq',
-                             metavar='SEQ',
-                             type=str,
-                             default='',
-                             help='Target ligand sequence if peptide, DNA, or RNA')
-    run_info.add_argument('-pk',
-                             '--pick_up',
-                             metavar='PICKUP',
-                             type=str,
-                             default='No',
-                             help='If the run is to resume an unfinished run or elongate another run.',
-                             choices=['Yes', 'No'])
-    
-    
[…]
