+++ main.py
@@ -24,10 +24,11 @@ import glob
 import os
 import platform
 import shutil
+from pathlib import Path
+import yaml
 from opendna import *
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
-
 
 
 parser = argparse.ArgumentParser(
@@ -106,17 +107,11 @@ paths.add_argument('-mb',
 # =========================== Parameters to run pipeline ===========================
 # "required" is set to be False as default, so that it is possible to only set it up in yaml config file
 
-run_info.add_argument('--test_mode',
+run_info.add_argument('--quick_check_mode',
                          metavar='\b',
                          type=str,
                          default='Yes',
-                         help='Rapidly run through pipeline using default test parameters',
-                         choices=['Yes','No'])
-run_info.add_argument('--explicit_run_enumeration',
-                         metavar='\b',
-                         type=str,
-                         default='Yes',
-                         help="'Yes': start a new run_num; 'No': output will be written to the an existing run_num folder.",
+                         help='Rapidly run a certain mode for quick check using default test parameters',
                          choices=['Yes','No'])
 run_info.add_argument('-r',
                          '--run_num',
@@ -125,7 +120,7 @@ run_info.add_argument('-r',
                          default=1,
                          help='Run number. Output will be written to {--workdir}/run{--run_num}/')
 
-
+# ================ DNA aptamer and ligand info ================
 run_info.add_argument('-m',
                          '--mode',
                          metavar='\b',
@@ -144,7 +139,6 @@ run_info.add_argument('-m',
     'full dock': 'free aptamer' + docking
     'full binding': 'full dock' + binding
 '''
-
 run_info.add_argument('-a',
                          '--aptamer_seq',
                          metavar='\b',
@@ -182,7 +176,7 @@ run_info.add_argument('--example_peptide_seq',
                          help='The sequence of the example peptide ligand')
 
 
-
+# ================ Pipeline module: NUPACK and MMB ================
 # # NUPACK and MMB modules:
 run_info.add_argument('--skip_MMB',
                          metavar='\b',
@@ -246,8 +240,7 @@ run_info.add_argument('--mmb_params',
                          help='Path to parameter file bundled with MMB package')
 
 
-
-
+# ================ Pipeline module: OpenMM ================
 # OpenMM module: 
 # # resume a MD sampling
 run_info.add_argument('-pk',
@@ -268,32 +261,19 @@ run_info.add_argument('--pickup_from_complexChk',
                          type=str,
                          default='No',
                          help='Resume MD sampling of aptamer-ligand: skip everything before',
-                         choices=['Yes','No']) # Not complete
+                         choices=['Yes','No'])
 run_info.add_argument('--chk_file',
                          metavar='\b',
                          type=str,
                          default=None,
-                         help='Name of checkpoint file for resuming MD sampling (file is placed in the current run{--run_num}')
+                         help='Name of checkpoint file for resuming MD sampling (path/filename.chk)')
 run_info.add_argument('--pickup_pdb',
                          metavar='\b',
                          type=str,
                          default=None,
-                         help='PDB file (topology + coordinates) for resuming MD sampling in explicit solvent (file is placed in the current run{--run_num}')
-# not complete: resume from top and crd file in implicit solvent
-# In opendna.py:
-# outputDict['free aptamer results {}'.format(self.i)] = self.freeAptamerDynamics(self.params['pickup_pdb'], implicitSolvent=self.params['implicit_solvent'])
-run_info.add_argument('--pickup_top',
[…]
