+++ main.py
@@ -1,11 +1,11 @@
+#!/usr/bin/env python
+
 '''
 E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
 
 An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
 
-Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine
-
-Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
+Copyright 2022 Michael Kilgour, Tao Liu and Lena Simine
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -19,41 +19,119 @@ Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
    See the License for the specific language governing permissions and
    limitations under the License.
 '''
+import argparse
+import glob
+import os
+import platform
+import shutil
+from pathlib import Path
+import yaml
 from opendna import *
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
-params = {}
-# ============================================= Specify your settings within this block for a local test ===================================================
-params['device'] = 'local'           # 'local' or 'cluster'
-params['device platform'] = 'macos'  # 'macos' or 'linux' or 'WSL' (Windows Subsystem for Linux). Not supporting pure Windows OS (due to NUPACK)
-params['platform'] = 'CPU'           # 'CPU' or 'CUDA'
-if params['platform'] == 'CUDA': params['platform precision'] = 'single'  # 'single' or 'double'
-if params['device'] == 'local':
-    params['workdir'] = '/path-to-E2EDNA2-JOSS/localruns'                  # directory manually created to store all future jobs
-    params['mmb dir'] = '/path-to-E2EDNA2-JOSS/Installer.3_0.OSX/lib'      # path to MMB dylib files  # need to tell OS where to find the library files. All MMB files are in the same direcotory.
-    params['mmb']     = '/path-to-E2EDNA2-JOSS/Installer.3_0.OSX/bin/MMB'  # path to the MMB executable; MMB is the *name* of the executable here
-else:  # params['device'] == 'cluster':
-    params['workdir'] = '/home/taoliu/scratch/runs'
-    params['mmb dir'] = '~/projects/def-simine/programs/MMB/Installer.2_14.Linux64'
-    params['mmb'] = '~/projects/def-simine/programs/MMB/Installer.2_14.Linux64/MMB.2_14.Linux64'
-params['explicit run enumeration'] = True  # To resume a previous run from .chk file, use ``False`` here
-cmdLineInputs = get_input()  # get input arguments from command lines
-params['run num'] = cmdLineInputs[0]
-params['mode'] = cmdLineInputs[1]
-params['aptamerSeq'] = cmdLineInputs[2]
-params['target ligand'] = cmdLineInputs[3]
-params['target ligand type'] = cmdLineInputs[4]
-params['target sequence'] = cmdLineInputs[5]
-# params['run num'] = 1  # 0: auto-increase run-num for a fresh run; > 0 AND params['explicit run enumeration'] = True: fresh run; > 0 AND params['explicit run enumeration'] = False: pickup on a previous run;
-# params['mode'] = 'full binding'  # specify simulation mode
-# params['aptamerSeq'] = 'TAATGTTAATTG'  # manually set DNA aptamer sequence from 5' to 3'
-# params['target ligand'] = 'YQTQ.pdb'  # pdb filename of the target ligand. If no target, use 'False', such as in 'free aptamer' mode.
[…]
