+++ opendna.py
@@ -1,5 +1,6 @@
 import sys
 import glob
+import shutil
 from shutil import copyfile, copytree
 
 import interfaces
@@ -13,41 +14,53 @@ class opendna:
     def __init__(self, params):
         self.workDir = ""  # rather use an empty string, not an empty list
         self.params = params
-        self.aptamerSeq = self.params['aptamerSeq']        
-        self.targetPDB = self.params['target ligand']        # either pdb filename (eg, "peptide.pdb") or boolean ``False``
-        self.targetType = self.params['target ligand type']  # 'peptide' or 'DNA' or 'RNA' or 'other'. Assume now the target is a peptide
-        self.targetSeq = self.params['target sequence']
-        # empty string, if target is not a peptide/DNA/RNA: how to deal with the "analyzeBinding"? No analysis in this simulator.
+        self.aptamerSeq = self.params['aptamer_seq']        
+        self.targetPDB = self.params['ligand']        # either pdb filename (eg, "/path/peptide.pdb") or None
+        self.targetType = self.params['ligand_type']  # 'peptide' or 'DNA' or 'RNA' or 'other' or None. Assume now the target is a peptide
+        self.targetSeq = self.params['ligand_seq']    # if target is not a peptide/DNA/RNA, targetSeq is None
+        #analyzeBinding method is commented out: no analysis in this simulator.
         
         self.pdbDict = {}
         self.dcdDict = {}
         self.actionDict = {}
         self.getActionDict()  # specify the actions based on the selected mode
-        # modify actions based on three toggles.
-        if self.params['skip MMB'] is True:  
-            self.actionDict['do 2d analysis'] = False  
-            self.actionDict['do MMB'] = False        
-        if self.params['pick up from freeAptamerChk'] is True:
-            # everything before "freeAptamerDynamics" is skipped
-            self.actionDict['do 2d analysis'] = False
-            self.actionDict['do MMB'] = False            
-            self.actionDict['do smoothing'] = False
-            self.actionDict['get equil repStructure'] = True  # making sure to run "freeAptamerDynamics"
-            # # Not used for now:
-            # printRecord('Copying the checkpoint file: {} to workdir.'.format(self.params['chk file']), self.workDir+'/')
-            # if os.path.exists(self.params['chk file']):
-            #     copyfile(self.params['chk file'], self.workDir + '/' + self.params['chk file'])
-            # else:
-            #     printRecord('Designated checkpoint file does not exist! Terminating the pipeline.', self.workDir+'/')
-            #     self.terminateRun()
-        elif self.params['pick up from complexChk'] is True:
-            # everything before "complexDynamics" is skipped
-            self.actionDict['do 2d analysis'] = False
-            self.actionDict['do MMB'] = False            
-            self.actionDict['do smoothing'] = False            
-            self.actionDict['get equil repStructure'] = False
-            self.actionDict['do docking'] = False
-            self.actionDict['do binding'] = True  # making sure to run "complexDynamics"
+        
+        # modify actions if necessary
+        if self.params['skip_MMB'] is True:  
+            self.actionDict['do MMB'] = False
+            if self.params['mode'] != '2d structure':
+                self.actionDict['do 2d analysis'] = False  
+        
+        if self.params['pickup'] is True:
+            printRecord('Copying the checkpoint file: {} to run folder: {}.'.format(self.params['chk_file'],self.workDir), self.workDir+'/')
+            printRecord('Copying the topology file: {} to run folder: {}.'.format(self.params['pickup_pdb'],self.workDir), self.workDir+'/')
+
+            # These conditions actually have been checked in main.py, here is for redundancy
+            if os.path.exists(self.params['chk_file']) and os.path.exists(self.params['pickup_pdb']):
[…]
