### 2022-05-11 siminegroup/E2EDNA2#14 @taoliu032
@JoaoRodrigues @rvhonorato Thank you both very much for the very helpful suggestions! I updated the codebase ([here](https://github.com/siminegroup/E2EDNA2/tree/beta_test)) accordingly and tested them myself already. Now we are waiting for any feedback from our beta tester. Once it is done (should not be long), I will update the JOSS branch and invite you to test the code.

==================
Summary of the updates:
- I implemented both `.yaml` configuration file and `argparse` to improve the way of handling the input arguments. Now, a user can edit input arguments via `.yaml` file or command line without touching source code.
  - I chose to provide both ways to increase the adaptability of the code. Sometimes, a user may want to iterate a list of possible values for an argument, then it might be easier to do so in command line rather than in a configuration file.
  - Currently, if an argument is specified in both `.yaml` and command line, the one in `.yaml` will overwrite that in command line. Such superiority of `.yaml` can be removed by commenting out that argument in `.yaml`.

- I reorganized the output directory (eg, `run2`) so that only the main outputs remain in the main output folder and all intermediate or temporary files are grouped to subfolders based on which module they were generated from. For example, `run2/mmb_temp_files_0/` stores temporary files generated by `MMB` and the `_0` indicates the index of a predicted secondary structure of DNA.

- In `examples/` folder, example `.yaml` files and `.sh` scripts for 2 sets of automated tests are provided, as well as a ligand structure (`example_peptide_ligand.pdb`) and pre-folded DNA structure (`foldedAptamer_0.pdb`).
  - To run a set of the automated tests:  simply copy a pair of `.sh` and `.yaml`  to the main codebase directory (contains `main.py`) then run `.sh` there.
  - One set of automated tests assumes the third-party dependencies `NUPACK` and `MMB` are not available: `automated_tests_no_NUPACK_MMB.sh` and `simu_config_automated_tests_no_NUPACK_MMB.yaml`

Thank you very much again.
