### 2022-05-11 #4182 @taoliu032
Hi @csoneson, thank you for your patience. I updated the codebase with the changes requested by the reviewers and tested them myself. The reviewers' suggestions are very helpful - thank you very much!

Now we are waiting for any feedback from our beta tester. Once it is done (should not be long), I will update the JOSS branch and invite reviewers to test the code, as the final step hopefully.

Summary of the updates can be seen here: https://github.com/siminegroup/E2EDNA2/issues/14#issuecomment-1124250089.
