+++ main.py
@@ -47,7 +47,7 @@ parser.add_argument('-yaml',
                          type=str,
                          default='simu_config.yaml',
                          required=True,
-                         help='A YAML configuration file that can specify all the arguments.')
+                         help='A YAML configuration file that can specify all the arguments')
 parser.add_argument('-ow',
                          '--overwrite',
                          action='store_true',
@@ -120,7 +120,7 @@ run_info.add_argument('-r',
                          metavar='\b',
                          type=int,
                          default=1,
-                         help='Run number. Output will be written to {--workdir}/run{--run_num}/')
+                         help='Run number. Output will be written to {--workdir}/run{--run_num}')
 
 # ================ DNA aptamer and ligand info ================
 run_info.add_argument('-m',
@@ -184,7 +184,7 @@ run_info.add_argument('--skip_MMB',
                          metavar='\b',
                          type=str,
                          default='No',
-                         help="'Yes': skip both 2D structure analysis and MMB folding, and start with a known --init_structure",
+                         help="If `Yes`: skip both 2D structure analysis and MMB folding, and start with a known --init_structure",
                          choices=['Yes','No'])
 run_info.add_argument('-init',
                          '--init_structure',
@@ -207,13 +207,13 @@ run_info.add_argument('--Mg_conc',
                          metavar='\b',
                          type=float,
                          default=0.0,
-                         help='Magnesium molar concentration used in NUPACK: [0, 0.2], default = 0.0')
+                         help='Magnesium molar concentration used in NUPACK: [0, 0.2]')
 
 run_info.add_argument('--fold_fidelity',
                          metavar='\b',
                          type=float,
                          default=0.9,
-                         help="Refold in MMB if score < fidelity (default = 0.9) unless the fold speed is 'quick'")
+                         help="Refold in MMB if score < `fold_fidelity` unless the `fold_speed` is `quick`")
 run_info.add_argument('--fold_speed',
                          metavar='\b',
                          type=str,
@@ -250,81 +250,81 @@ run_info.add_argument('-pk',
                          metavar='\b',
                          type=str,
                          default='No',
-                         help='Whether the run is to resume MD sampling of an unfinished run or an old run.',
+                         help='Whether the run is to resume MD sampling of an unfinished run or an old run',
                          choices=['Yes', 'No'])
 run_info.add_argument('--pickup_from_freeAptamerChk',
                          metavar='\b',
                          type=str,
                          default='No',
-                         help='Resume MD sampling of free aptamer: skip everything before',
+                         help='Resume MD sampling of free aptamer: skip everything before it',
                          choices=['Yes','No'])
 run_info.add_argument('--pickup_from_complexChk',
                          metavar='\b',
                          type=str,
                          default='No',
-                         help='Resume MD sampling of aptamer-ligand: skip everything before',
+                         help='Resume MD sampling of aptamer-ligand: skip everything before it',
                          choices=['Yes','No'])
 run_info.add_argument('--chk_file',
                          metavar='\b',
                          type=str,
                          default=None,
-                         help='Name of checkpoint file for resuming MD sampling (path/filename.chk)')
+                         help='Name of checkpoint file for resuming MD sampling, format: <path>/<filename>.chk')
 run_info.add_argument('--pickup_pdb',
                          metavar='\b',
                          type=str,
                          default=None,
-                         help='PDB file (topology+coordinates) for resuming MD sampling in explicit solvent (path/filename.pdb')
+                         help='PDB file (topology+coordinates) for resuming MD sampling in explicit solvent, format: <path>/<filename>.pdb')
 
 # # Environmental parameters
 run_info.add_argument('--pressure',
                          metavar='\b',
                          type=float,
                          default=1.0,
[…]
