### 2022-05-12 1d7e3a1
Update the JOSS branch code (#16)
* Add files via upload

* Delete lib/peptide directory

* Delete YQTQ.pdb

* Delete automate_tests.sh

* Delete foldedAptamer_0.pdb

* Delete backbone_dihedrals.csv

* Update codebase

* Upload examples and scripts for automated tests

* Delete LICENSE

* Update automated tests

* Update README

* Update README.MD
