+++ utils.py
@@ -1,5 +1,5 @@
 """
-Utitilies -- how to intuitively distinguish it from analysisTools.py?
+Utitilies
 """
 from openmm.app import *
 import openmm.unit as unit
@@ -19,34 +19,6 @@ from collections import Counter
 
 from PeptideBuilder import Geometry  # pip install PeptideBuilder
 import PeptideBuilder
-
-
-# I/O
-def get_input():
-    """
-    get the command line in put for the run num. defaulting to a new run (0)
-    :return:
-    """
-    # Two examples of commands:
-    # python main.py --run_num=1 --mode='free aptamer' --aptamerSeq='TAATGTTAATTG' --ligand='False' --ligandType='' --ligandSeq=''
-    # python main.py --run_num=2 --mode='full dock' --aptamerSeq='TAATGTTAATTG' --ligand='YQTQ.pdb' --ligandType='peptide' --ligandSeq='YQTQTNSPRRAR'
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--run_num', type=int, default=0)
-    parser.add_argument('--mode', type=str, default='simulation_mode')
-    parser.add_argument('--aptamerSeq', type=str, default='NOSEQUENCE')
-    parser.add_argument('--ligand', type=str, default='False')
-    parser.add_argument('--ligandType', type=str, default='')
-    parser.add_argument('--ligandSeq', type=str, default='')
-    
-    cmd_line_input = parser.parse_args()
-    run_num = cmd_line_input.run_num
-    mode = cmd_line_input.mode
-    aptamerSeq = cmd_line_input.aptamerSeq
-    ligand = cmd_line_input.ligand
-    ligandType = cmd_line_input.ligandType
-    ligandSeq = cmd_line_input.ligandSeq
-    
-    return [run_num, mode, aptamerSeq, ligand, ligandType, ligandSeq]
 
 
 # def recenterDCD(topology, trajectory):
@@ -80,11 +52,11 @@ def printRecord(statement, filepath=""):
     :return:
     """
     print(statement)
-    if os.path.exists(filepath + 'record.txt'):
-        with open(filepath + 'record.txt', 'a') as file:
+    if os.path.exists(filepath + 'run_output_log.txt'):
+        with open(filepath + 'run_output_log.txt', 'a') as file:
             file.write('\n' + statement)
     else:
-        with open(filepath + 'record.txt', 'w') as file:
+        with open(filepath + 'run_output_log.txt', 'w') as file:
             file.write('\n' + statement)
 
 
@@ -208,11 +180,21 @@ def cleanTrajectory(structure, trajectory):
     :param trajectory:
     :return:
     """
+    structureName  = os.path.basename(structure)
+    trajectoryName = os.path.basename(trajectory)
+    structureDir  = os.path.dirname(structure)
+    trajectoryDir = os.path.dirname(trajectory)
+    
+    # cleaned files are in the same directory as the uncleaned ones.
+    clean_structure  = os.path.join(structureDir, 'clean_'+structureName)
+    clean_trajectory = os.path.join(trajectoryDir, 'clean_'+trajectoryName)
+
     u = mda.Universe(structure, trajectory)
     # TODO: if u.segments.n_segments > 2:  # if > 2 segments, then there must be solvent and salts (assuming nonzero salt concentration)
     goodStuff = u.segments[:-2].atoms  # cut out salts and solvents
-    goodStuff.write("clean_" + structure)  # write topology
-    with mda.Writer("clean_" + trajectory, goodStuff.n_atoms) as W:
+
+    goodStuff.write(clean_structure)  # write topology
+    with mda.Writer(clean_trajectory, goodStuff.n_atoms) as W:
         for ts in u.trajectory:  # indexing over the trajectory
             W.write(goodStuff)
 
[…]
