+++ e2edna-env.yml
@@ -174,4 +174,5 @@ dependencies:
     - pytraj==2.0.6
     - sander==16.0
     - seqfold==0.7.11
+    - pyyaml==6.0
 

[…]
