### 2022-05-13 #4182 @csoneson
@taoliu032 I have made a [PR](https://github.com/siminegroup/E2EDNA2/pull/17) with a couple of minor edits to the manuscript - please check and merge if you agree. After merging, you can generate a new proof with `@editorialbot generate pdf` as above.

The next steps then are as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
