### 2022-05-13 de904d9
Merge argument parsing branch (#18)
* add shebang

* add argument parsing

* begin validating arguments

* fix small inconsistencies

* update argument formatting

* improve directory handling

* automatically detect OS

* revert to run_num as argument name

* update readme to reflect argparse implementation

* revert to --run_num argument name

* broaden default mmb location flag wildcards

* fix a few typos

* remove accidental extra characters

* remove old command line parsing

* use argparse to validate --mode

=== The commits above were made by schackartk and were merged into the topic b
[…]
