### 2022-05-13 b385138
Minor changes (#17)
* Minor modifications

* A couple more minor edits
