+++ paper/paper.bib
@@ -247,7 +247,7 @@
 	number = {7},
 	urldate = {2022-01-27},
 	journal = {Bioinformatics},
-	author = {Caroli, Jimmy and Forcato, Mattia and Bicciato, Silvio},
+	author = {Caroli, J. and Forcato, Mattia and Bicciato, Silvio},
 	editor = {Berger, Bonnie},
 	month = apr,
 	year = {2020},
@@ -267,7 +267,7 @@
 	number = {14},
 	urldate = {2022-01-27},
 	journal = {Current Pharmaceutical Design},
-	author = {O. Tucker, W. and T. Shum, K. and A. Tanner, J.},
+	author = {Tucker, W. O. and Shum, K. T. and Tanner, J. A.},
 	month = mar,
 	year = {2012},
 	pages = {2014--2026},
@@ -298,8 +298,12 @@
 	abstract = {Aptamers, single stranded nucleic acid-based molecular recognition elements, are generated through an in vitro evolution process called SELEX. The selection of aptamers with optimal specificity and high affinity remains a challenge, leading to the recent emergence of elegant new approaches and tools for SELEX. In this commentary, we discuss recent advancements in the selection of aptamers.},
 	language = {en},
 	author = {McKeague, Maureen and DeRosa, Maria C},
-	pages = {5},
+	pages = {12-16},
 	file = {McKeague and DeRosa - Aptamers and SELEX Tools for the Development of T.pdf:/Users/taoliu/Zotero/storage/NKQAGNZL/McKeague and DeRosa - Aptamers and SELEX Tools for the Development of T.pdf:application/pdf},
+        journal = {Aptamers and Synthetic Antibodies},
+        year = {2014},
+        volume = {1},
+        number = {1} 
 }
 
 @article{Tuerk:1990,
