### 2022-05-13 390a7ad
A minor change in paper.md as mentioned in #17
