### 2022-05-15 c78b7eb
Squashed commit from beta_test to main branch
- Keep these folders in main branch: Module_Readme, old, and testScripts
- Exclude the folder paper from beta_test branch
