+++ interfaces.py
@@ -6,10 +6,10 @@ OpenMM (omm)
 LightDock (ld)
 NUPACK
 """
-
+import sys
+import shutil
 from shutil import copyfile
 from numpy import pi
-from nupack import *
 
 from openmm import *
 from openmm.app import *
@@ -31,7 +31,7 @@ class nupack:
         self.mgConc = mgConc
         self.R = 0.0019872  # ideal gas constant in kcal/mol/K
 
-    def run(self):
+    def run(self):        
         self.energyCalc()
         self.structureAnalysis()
         return self.ssDict
@@ -44,6 +44,13 @@ class nupack:
         Ouput a lot of analysis results in self.output
         :return:
         """
+        try:
+            import nupack
+            # from nupack import * #: does not work at this level
+        except ImportError as err:
+            raise ImportError("Required module 'nupack' not found. Please download from http://www.nupack.org/downloads and install in your Python virtual environment.") from err            
+        
+        # Run nupack-dependent code
         if self.temperature > 273:  # auto-detect Kelvins
             gap = 2 * self.R * self.temperature
             CelsiusTemprature = self.temperature - 273
@@ -51,13 +58,12 @@ class nupack:
             gap = 2 * self.R * (self.temperature + 273)  # convert to Kelvin fir kT
             CelsiusTemprature = self.temperature
 
-        A = Strand(self.aptamerSeq, name='A')
-        comp = Complex([A], name='AA')
-        set1 = ComplexSet(strands=[A], complexes=SetSpec(max_size=1, include=[comp]))
-        model1 = Model(material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
-        results = complex_analysis(set1, model=model1, compute=['pfunc', 'mfe', 'subopt', 'pairs'], options={'energy_gap': gap})
+        A = nupack.Strand(self.aptamerSeq, name='A')
+        comp = nupack.Complex([A], name='AA')
+        set1 = nupack.ComplexSet(strands=[A], complexes=nupack.SetSpec(max_size=1, include=[comp]))
+        model1 = nupack.Model(material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
+        results = nupack.complex_analysis(set1, model=model1, compute=['pfunc', 'mfe', 'subopt', 'pairs'], options={'energy_gap': gap})
         self.output = results[comp]
-    # TODO: what do these nupack functions do?
 
     def structureAnalysis(self):
         """
@@ -98,14 +104,14 @@ class nupack:
 
 class mmb:  # MacroMolecule Builder (MMB)
     def __init__(self, aptamerSeq, pairList, params, ind1, intervalLength=None):
-        if params['fold speed'] == 'quick':
+        if params['fold_speed'] == 'quick':
             self.template = 'commands.template_quick.dat'  # can be used for debugging runs - very short
-        elif params['fold speed'] == 'normal':
+        elif params['fold_speed'] == 'normal':
             self.template = 'commands.template.dat'        # default folding algorithm
-        elif params['fold speed'] == 'long':
+        elif params['fold_speed'] == 'slow':
             self.template = 'commands.template_long.dat'   # extended annealing - for difficult sequences
         self.comFile = 'commands.run_fold.dat'
-        self.foldSpeed = params['fold speed']
+        self.foldSpeed = params['fold_speed']
         self.aptamerSeq = aptamerSeq
         self.temperature = params['temperature']
         self.pairList = pairList
@@ -113,13 +119,13 @@ class mmb:  # MacroMolecule Builder (MMB)
         
         # Conditions used to decide how to run MMB executable
         self.device = params['device']  # 'local' or 'cluster'
[…]
