### 2022-05-15 762ec4b
Delete automate_tests.sh
Replaced by the scripts in examples/automated_tests/
