### 2022-05-16 c7adff8
Update the license to GPLv3
