+++ analysisTools.py
@@ -1,3 +1,24 @@
+'''
+E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
+
+An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
+    
+Copyright (C) 2022 Michael Kilgour, Tao Liu and Lena Simine
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+'''
+
 # tools for trajectory analysis
 from utils import printRecord
 import numpy as np
