### 2022-05-16 #4182 @taoliu032
Hi @csoneson! I believe I have completed the following steps.
> * [x]  Make a tagged release of the software, and list the version tag of the archived version here.

The release is [here](https://github.com/siminegroup/E2EDNA2/releases/tag/v2.0.0) and the version tag is `v2.0.0`, in order to be explicitly distinguished from the original version of E2EDNA.

We would like to ask a question regarding the version: When we made our initial submission to JOSS, we were asked to provide a "Software version". We first tried "v2.0.0" but the submission portal did not allow us to do so, probably because this is our very first paper in JOSS about this software and we chose "New submission" as "Type of submission". So we ended up putting down "v1.0", which is now shown at the top of this review issue and quoted below:
>> **Submitting author:** @taoliu032 ([Tao Liu](http://orcid.org/0000-0002-1082-5570))
> **Repository:** https://github.com/siminegroup/E2EDNA2
> **Branch with paper.md** (empty if default branch): JOSS
> **Version:** v1.0

Is it okay to have such a discrepancy, submission version "v1.0" v.s. release version "v2.0.0"? Where would the "v1.0" and "v2.0.0" show up later?

> * [x]  Archive the reviewed software in Zenodo or a similar service
> * [x]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.

The deposit was created in Zenodo and metadata was checked: https://zenodo.org/record/6546661

> * [x]  Please list the DOI of the archived version here.

DOI on Zenodo: 10.5281/zenodo.6546661

Thank you again!
