### 2022-05-18 #4182 @arfon
/ cc @tarleb – any thoughts on how we should handle this? It looks like the BibTeX file has two ISSNs.
