### 2022-05-18 #4182 @tarleb
I've pushed a change, now only the first ISSN will be used.

Publications can have an ISSN for their print version and another for the electronic (online) version. The latter is sometimes called the eISSN and it seems that Zotero adds both.

There may be other ways to improve our handling of the ISSN field, but it should work for now.
