### 2022-05-21 06e90f6
Incorporate major updates from E2EDNA 2.0.0
-use new ways of passing input arguments: argparse and yaml
- update the .py files with new `params` dict variable
- delete obsolete files
- reorganize the output directory: group temporary files to subfolders
- fix a bug in opendna.py of E2EDNA 2.0.0: if starting with a folded structure whose filename is not foldedAptamer_0.pdb, will continue to use the original filename for MD or docking.
