+++ main.py
@@ -1,60 +1,139 @@
+#!/usr/bin/env python
+
 '''
 E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
 
 An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
+    
+Copyright (C) 2022 Michael Kilgour, Tao Liu and Lena Simine
 
-Michael Kilgour, Tao Liu, Ilya S. Dementyev, Lena Simine
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
 
-Copyright 2022 Michael Kilgour, Tao Liu and contributors of E2EDNA 2.0
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
 
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
 '''
+
+import argparse
+import glob
+import os
+import platform
+import shutil
+from pathlib import Path
+import yaml
 from opendna import *
 import warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
-params = {}
-# ============================================= Specify your settings within this block for a local test ===================================================
-params['device'] = 'cluster'           # 'local' or 'cluster'
-params['device platform'] = 'linux'  # 'macos' or 'linux' or 'WSL' (Windows Subsystem for Linux). Not supporting pure Windows OS (due to NUPACK)
-params['platform'] = 'CUDA'           # 'CPU' or 'CUDA'
-if params['platform'] == 'CUDA': params['platform precision'] = 'single'  # 'single' or 'double'
-if params['device'] == 'local':
-    params['workdir'] = '/Users/taoliu/Downloads/E2EDNA2-JOSS/localruns'                  # directory manually created to store all future jobs
-    params['mmb dir'] = '/Users/taoliu/Downloads/E2EDNA2-JOSS/Installer.3_0.OSX/lib'      # path to MMB dylib files  # need to tell OS where to find the library files. All MMB files are in the same direcotory.
-    params['mmb']     = '/Users/taoliu/Downloads/E2EDNA2-JOSS/Installer.3_0.OSX/bin/MMB'  # path to the MMB executable; MMB is the *name* of the executable here
-else:  # params['device'] == 'cluster':   
-    params['workdir'] = '/home/taoliu/scratch/adenosine_binding_aptamer/27nt_aptamer/test_amoeba_FF/runs/freeAptamer'
-    params['mmb dir'] = '~/projects/def-simine/programs/MMB/Installer.2_14.Linux64'
-    params['mmb'] = '~/projects/def-simine/programs/MMB/Installer.2_14.Linux64/MMB.2_14.Linux64'
-params['explicit run enumeration'] = True  # To resume a previous run from .chk file, use ``False`` here
[…]
