+++ interfaces.py
@@ -1,3 +1,24 @@
+'''
+E2EDNA 2.0 - OpenMM Implementation of E2EDNA !
+
+An automated pipeline for simulating DNA aptamers complexed with target ligands (peptide, DNA, RNA or small molecules).
+    
+Copyright (C) 2022 Michael Kilgour, Tao Liu and Lena Simine
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+'''
+
 """
 interfaces for other major software used by E2EDNA
 MacroMolecule Builder (MMB)
@@ -6,10 +27,10 @@ OpenMM (omm)
 LightDock (ld)
 NUPACK
 """
-
+import sys
+import shutil
 from shutil import copyfile
 from numpy import pi
-from nupack import *
 
 from openmm import *
 from openmm.app import *
@@ -22,12 +43,12 @@ from analysisTools import *
 
 
 class nupack:
-    def __init__(self, aptamerSeq, temperature, ionicStrength, mgConc=0, stacking='nostacking'):
+    def __init__(self, aptamerSeq, temperature, ionicStrength, mgConc=0, ensemble='nostacking'):
         self.aptamerSeq = aptamerSeq
         self.temperature = temperature
         self.naConc = ionicStrength
         self.mgConc = mgConc
-        self.stacking = stacking
+        self.ensemble = ensemble
         self.R = 0.0019872  # ideal gas constant in kcal/mol/K
 
     def run(self):
@@ -43,7 +64,13 @@ class nupack:
         Ouput a lot of analysis results in self.output
         :return:
         """
-        # Compute free energy gap used by looking for suboptimal structure whose energy <= delta_G_mfe + gap
+        try:
+            import nupack
+            # from nupack import * #: does not work at this level
+        except ImportError as err:
+            raise ImportError("Required module 'nupack' not found. Please download from http://www.nupack.org/downloads and install in your Python virtual environment.") from err            
+        
+        # Run nupack-dependent code
         if self.temperature > 273.15:  # auto-detect Kelvins
             gap = 2 * self.R * self.temperature
             CelsiusTemprature = self.temperature - 273.15
@@ -51,14 +78,13 @@ class nupack:
             gap = 2 * self.R * (self.temperature + 273.15)  # convert to Kelvin for kT
             CelsiusTemprature = self.temperature
 
-        A = Strand(self.aptamerSeq, name='A') # specify a strand and name
+        A = nupack.Strand(self.aptamerSeq, name='A') # specify a strand and name
         # A.nt() # calculate the number of nucleotides
-        comp = Complex([A], name='AA', bonus=0) # specify a complex of one or more interacting strands and name is optional.
-        
-        # specify a set of strands that interact to form a set of complexes
-        set1 = ComplexSet(strands=[A], complexes=SetSpec(max_size=1, include=[comp])) # 1 complex set: [[A]]
+        comp = nupack.Complex([A], name='AA', bonus=0) # specify a complex of one or more interacting strands and name is optional.
 
-        model1 = Model(ensemble=self.stacking, material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
[…]

[…]
