### 2022-05-21 2aa8f0b
Remove old/ and testScripts/
Remove them so that they won't get copied to a new branch when created. If somehow need to retrieve them, check out this commit.
