+++ testScripts/lightdockConvergenceTest.py
@@ -1,167 +1,0 @@
-import os
-import numpy as np
-from utils import *
-from analysisTools import *
-import tqdm
-import time
-
-'''
-script to test convergence of lighdock docking with a small peptide and medium sized aptamer
-currently we use swarms = 1 (scaling factor meaning 1x the calculated required number of swarms for a given surface area)
-'''
-
-# get relevant paths
-
-
-params = {}
-params['device'] = 'local'  # 'local' or 'cluster'
-params['N docked structures'] = 10  # number of docked structures to output from the docker. If running binding, it will go this time (at linear cost)
-params['pH'] = 7.4
-# paths
-if params['device'] == 'local':
-    params['workdir'] = '/home/mkilgour/mmruns' #'/mnt/c/Users/mikem/Desktop/mmruns'
-    params['mmb dir'] = '/mnt/c/Users/mikem/Desktop/software/Installer.2_14.Linux64'
-    params['mmb'] = '/mnt/c/Users/mikem/Desktop/software/Installer.2_14.Linux64/MMB.2_14.Linux64'
-    # lightdock python scripts
-    params['ld setup path'] = 'ld_scripts/lightdock3_setup.py'
-    params['ld run path'] = 'ld_scripts/lightdock3.py'
-    params['lgd generate path'] = '../ld_scripts/lgd_generate_conformations.py'
-    params['lgd cluster path'] = '../ld_scripts/lgd_cluster_bsas.py'
-    params['lg ant path'] = 'ld_scripts/ant_thony.py'
-    params['lgd rank path'] = 'ld_scripts/lgd_rank.py'
-    params['lgd top path'] = 'ld_scripts/lgd_top.py'
-
-elif params['device'] == 'cluster':
-    params['workdir'] = '/home/kilgourm/scratch/mmruns'  # specify your working directory here
-    params['mmb dir'] = '~/projects/def-simine/programs/MMB/Installer.2_14.Linux64'  # 'C:/Users/mikem/Desktop/Installer.2_14.Windows/MMB.2_14.exe'
-    params['mmb'] = '~/projects/def-simine/programs/MMB/Installer.2_14.Linux64/MMB.2_14.Linux64'
-    # lightdock python scripts
-    params['ld setup path'] = 'python ld_scripts/lightdock3_setup.py'
-    params['ld run path'] = 'python ld_scripts/lightdock3.py'
-    params['lgd generate path'] = 'python ../ld_scripts/lgd_generate_conformations.py'
-    params['lgd cluster path'] = 'python ../ld_scripts/lgd_cluster_bsas.py'
-    params['lg ant path'] = 'python ld_scripts/ant_thony.py'
-    params['lgd rank path'] = 'python ld_scripts/lgd_rank.py'
-    params['lgd top path'] = 'python ld_scripts/lgd_top.py'
-
-
-# initialize working directory
-os.chdir('/mnt/c/Users/mikem/Desktop/mmruns/lightdockConvergenceRuns/run1')
-
-# initialize docker
-
-class ld(): # lightdock
-    def __init__(self, aptamer, peptide, params, ind1):
-        self.setupPath = params['ld setup path']
-        self.runPath = params['ld run path']
-        self.genPath = params['lgd generate path']
-        self.clusterPath = params['lgd cluster path']
-        self.rankPath = params['lgd rank path']
-        self.topPath = params['lgd top path']
-
-        self.aptamerPDB = aptamer
-        self.peptidePDB = peptide
-
-        self.glowWorms = params['worms']
-        self.dockingSteps = params['docking steps']
-        self.numTopStructures = params['N docked structures']
-
-        self.pH = params['pH']
-
-        self.ind1 = ind1
-
-    def getSwarmCount(self):
-        # identify aptamer size
[…]
