+++ testScripts/secondaryStructureScripts.py
@@ -1,182 +1,0 @@
-from nupack import *
-from utils import *
-from seqfold import dg, fold
-
-
-def getSeqfoldStructure(sequence,temperature):
-    '''
-    output the secondary structure for a given sequence at a given condition
-    formats - ss string and pair list
-    '''
-    dg(sequence, temp=temperature)  # get energy of the structure
-    # print(round(sum(s.e for s in structs), 2)) # predicted energy of the final structure
-
-    structs = fold(sequence)  # identify structural features
-    desc = ["."] * len(sequence)
-    pairList = []
-    for s in structs:
-        pairList.append(s.ij[0])
-        pairList[-1]  # list of bound pairs indexed from 1
-        if len(s.ij) == 1:
-            i, j = s.ij[0]
-            desc[i] = "("
-            desc[j] = ")"
-
-    ssString = "".join(desc)
-    pairList = np.asarray(pairList) + 1
-
-    return ssString, pairList
-
-
-def getNupackString(sequence,temperature,ionicStrength):
-    R = 0.0019872 # ideal gas constant in kcal/mol/K
-    gap = R * temperature
-    A = Strand(sequence, name='A')
-    comp = Complex([A], name='AA')
-    set1 = ComplexSet(strands=[A], complexes=SetSpec(max_size=1, include=[comp]))
-    model1 = Model(material='dna', celsius = temperature - 273, sodium= ionicStrength)
-    results = complex_analysis(set1, model=model1, compute=['pfunc', 'mfe','subopt','pairs'], options={'energy_gap':gap})
-    cout = results[comp]
-    prob = np.exp(-cout.mfe[0].energy/R/temperature)/float(cout.pfunc) # probability - energy/partition function
-    ssString = cout.mfe[0].structure
-    # look at suboptimal structures
-    subopts = cout.subopt
-    subStrings, subProbs, subEns, subStack, subProbs = [[],[],[],[],[]]
-    for subopt in subopts:
-        subProbs.append(np.exp(-subopt.energy/R/temperature)/float(cout.pfunc))
-
-
-    return str(ssString), prob, subopts, subProbs
-
-
-def getNupackStructure(sequence,temperature,ionicStrength):
-    '''
-    compute secondary structure using NUPACK potentials
-    inputs - sequence, temperature in K, ionic strength (limited range of working ionic strengths)
-    '''
-    A = Strand(sequence, name='A')
-    comp = Complex([A], name='AA')
-    set1 = ComplexSet(strands=[A], complexes=SetSpec(max_size=1, include=[comp]))
-    model1 = Model(material='dna', celsius = temperature - 273, sodium= ionicStrength)
-    results = complex_analysis(set1, model=model1, compute=['pfunc', 'pairs', 'subopt', 'mfe'])
-    cout = results[comp]
-
-    '''
-    # print the secondary structure strings
-    print('\nMFE proxy structures(s) for set')
-    for i, s, in enumerate(cout.mfe):
-        print('    %2d: %s (%.2f kcal/mol)' % (i, s.structure, s.energy))
-
-    print('\nSuboptimal proxy structures for set')
-    for i, s in enumerate(cout.subopt):
-        print('    %2d: %s (%.2f kcal/mol)' % (i, s.structure, s.energy))
-    '''
-
-    # extract list of most probable base pairs
-    pairMat = cout.pairs.to_array()
-    pairList = []
-    paired = []
-    for i in range(len(pairMat)): # this whole thing doesn't seem to work properly to me
-        bestMatch = np.argmax(pairMat[i,:])
-        if True:#pairMat[i,bestMatch] > 0.5: # if we're relatively confident about this secondary structure feature
-            if bestMatch != i: # if the best match is itself, it means it does not bind
[…]
