### 2022-05-21 c93fbd8
Start the update from v2.0.0 to v2.1.0 (or v2.0.1)
- Add `ensemble` option in NUPACK module
- Delete `templates` directory
- Add two more yaml files for using amoeba FF or implicit solvent models: `simu_config_amoeba.yaml` and `simu_config_implicit_solvent.yaml`
- Bug fix in opendna.py
  - If starting with a folded structure, its filename will not be changed to "foldedAptamer_0.pdb" but be retained for the following steps: MD or docking.
- Correct typo in README
