+++ interfaces.py
@@ -35,24 +35,23 @@ from numpy import pi
 from openmm import *
 from openmm.app import *
 import openmm.unit as unit
+# In OpenMM/7.7.0: the prefix of "simtk" has been removed.
+# In OpenMM/7.5.0: package simtk is created with packages "openmm" and "unit" inside but literally import the stand-alone openmm.
 
 from utils import *
 from analysisTools import *
 
-# installed openmm: http://docs.openmm.org/7.5.0/userguide/application.html#installing-openmm
-# openmm is installed as an individual package;
-# package simtk is created with packages "openmm" and "unit" inside but literally import the stand-alone openmm.
-
 
 class nupack:
-    def __init__(self, aptamerSeq, temperature, ionicStrength, mgConc=0):
+    def __init__(self, aptamerSeq, temperature, ionicStrength, mgConc=0, ensemble='nostacking'):
         self.aptamerSeq = aptamerSeq
         self.temperature = temperature
         self.naConc = ionicStrength
         self.mgConc = mgConc
+        self.ensemble = ensemble
         self.R = 0.0019872  # ideal gas constant in kcal/mol/K
 
-    def run(self):        
+    def run(self):
         self.energyCalc()
         self.structureAnalysis()
         return self.ssDict
@@ -72,19 +71,51 @@ class nupack:
             raise ImportError("Required module 'nupack' not found. Please download from http://www.nupack.org/downloads and install in your Python virtual environment.") from err            
         
         # Run nupack-dependent code
-        if self.temperature > 273:  # auto-detect Kelvins
+        if self.temperature > 273.15:  # auto-detect Kelvins
             gap = 2 * self.R * self.temperature
-            CelsiusTemprature = self.temperature - 273
+            CelsiusTemprature = self.temperature - 273.15
         else:
-            gap = 2 * self.R * (self.temperature + 273)  # convert to Kelvin fir kT
+            gap = 2 * self.R * (self.temperature + 273.15)  # convert to Kelvin for kT
             CelsiusTemprature = self.temperature
 
-        A = nupack.Strand(self.aptamerSeq, name='A')
-        comp = nupack.Complex([A], name='AA')
-        set1 = nupack.ComplexSet(strands=[A], complexes=nupack.SetSpec(max_size=1, include=[comp]))
-        model1 = nupack.Model(material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
-        results = nupack.complex_analysis(set1, model=model1, compute=['pfunc', 'mfe', 'subopt', 'pairs'], options={'energy_gap': gap})
-        self.output = results[comp]
+        A = nupack.Strand(self.aptamerSeq, name='A') # specify a strand and name
+        # A.nt() # calculate the number of nucleotides
+        comp = nupack.Complex([A], name='AA', bonus=0) # specify a complex of one or more interacting strands and name is optional.
+
+        # specify a set of strands that interact to form a set of complexes
+        set1 = nupack.ComplexSet(strands=[A], complexes=nupack.SetSpec(max_size=1, include=[comp])) # 1 complex set: [[A]]
+        model1 = nupack.Model(ensemble=self.ensemble, material='dna', celsius=CelsiusTemprature, sodium=self.naConc, magnesium=self.mgConc)
+        '''
+        - ensemble: whether to consider coaxial or dangle stacking for multiloop or exterior loop. 
+            - 'stacking': default. Complex ensemble with coaxial and dangle stacking
+            - 'nostacking': Complex ensemble without coaxial and dangle stacking
+            - Historical options in nupack3 are supported too: all are about dangle stacking and no coaxial stacking.
+                - 'none-nupack3', 'some-nupack3', 'all-nupack3'
[…]
