+++ main.py
@@ -90,7 +90,7 @@ paths.add_argument('-w',
                     '--workdir',
                     metavar='DIR',
                     type=str,
-                    default='./localruns',
+                    default='./e2edna_runs',
                     help='Working directory to store individual output runs')
 paths.add_argument('-mbdir',
                     '--mmb_dir',
@@ -200,6 +200,12 @@ run_info.add_argument('--secondary_structure_engine',
                          type=str,
                          default='NUPACK',
                          help='Pipeline module that is used to predict secondary structures')
+run_info.add_argument('--ensemble',
+                         metavar='\b',
+                         type=str,
+                         default='nostacking',
+                         help='How NUPACK4 handles the stacking calculations',
+                         choices=['nostacking', 'stacking', 'none-nupack3', 'some-nupack3', 'all-nupack3'])
 run_info.add_argument('--N_2D_structures',
                          metavar='\b',
                          type=int,

[…]
