### 2022-05-23 #4182 @taoliu032
Hi @arfon! Thanks for getting us help to solve the ISSN issue. May I ask what steps are left for accepting the manuscript? Our editor recommended acceptance but the command to the robot may not completely go through, due to the ISSN issue.
