### 2022-05-30 #4182 @csoneson
Thanks @tarleb! Looks like the same fix may need to be applied to the ISBN field 🙂
