+++ paper/paper.md
@@ -28,13 +28,13 @@ bibliography: paper.bib
 
 DNA aptamers are short sequences of single-stranded DNA with untapped potential in molecular medicine, drug design, and materials design due to their strong and selective and most importantly tunable binding affinity to target molecules [@Tucker:2012; @Zhou:2017]. For instance, DNA aptamers can be used as therapeutics [@Corey:2021] for a wide range of diseases such as epilepsy [@Zamay:2020] and cancer [@Morita:2018]. They can also be used to detect a wide variety of molecular ligands, including antibiotics [@Mehlhorn:2018], neurotransmitters [@Sinha:2020], metals [@Qu:2016], proteins [@Kirby:2004], nucleotides [@Shen:2007] and metabolites [@Dale:2021; @Dauphin-Ducharme:2022] in real time, even in harsh environments [@McConnell:2020].
 
-We present E2EDNA 2.0: End-2-End DNA 2.0, a python simulation pipeline which offers a unified and automated solution to computational modeling of DNA aptamers with molecular ligands. It is broadly aimed at researchers developing therapeutics and sensors based on DNA aptamers who require detailed atomistic information on the behavior of aptamers and ligands in realistic media. Similar to its predecessor E2EDNA [@Kilgour:2021], E2EDNA 2.0 predicts DNA aptamers’ secondary and tertiary structures, and if a ligand is present,  the configuration of the solvated aptamer-ligand complex. 
+We present E2EDNA 2.0: End-2-End DNA 2.0, a Python simulation pipeline that offers a unified and automated solution to computational modeling of DNA aptamers with molecular ligands. It is broadly aimed at researchers developing therapeutics and sensors based on DNA aptamers who require detailed atomistic information on the behavior of aptamers and ligands in realistic media. Similar to its predecessor E2EDNA [@Kilgour:2021], E2EDNA 2.0 predicts DNA aptamers’ secondary and tertiary structures, and if a ligand is present,  the configuration of the solvated aptamer-ligand complex. 
 
 # Statement of Need
 
-With E2EDNA 2.0 our goal is to create a python-interfacing simulation package for single-stranded DNA with small ligands that is easy to install and use in python-based workflows. The pipeline is automated, yet flexible, taking us from DNA aptamer sequence to folded aptamer, and aptamer-ligand complex. Currently, available software packages designed for computationally studying DNA aptamers predominantly focus on partial feature analyses of RNA and DNA aptamers. For example, APTANI [@Caroli:2016] and APTANI2 [@Caroli:2020], commonly used by the aptamer community, select potentially relevant aptamers from SELEX (Systematic Evolution of Ligands by EXponential enrichment)[@Tuerk:1990; @McKeague:2014] experimental data sets through a sequence-structure analysis, and AEGIS, a platform equipped with a generative deep learning model to propose novel aptamer sequences [@Biondi:2018]. These approaches are aimed at fast black-box analysis of large numbers of sequences. A detailed analysis of a small number of high-promise candidate sequences is often required but no automated and easy-to-use simulation package exists in the computational space aside from E2EDNA [@Kilgour:2021] and E2EDNA 2.0, with the latter preferable for ease of installation and user-friendly implementations of eight modes of simulation.
[…]
