### 2022-05-30 #4182 @danielskatz
@taoliu032 - I'm the AEiC this week, and have a few small changes for the paper: please merge https://github.com/siminegroup/E2EDNA2/pull/20 or let me know what you disagree with, then we can proceed to publication.
